Clean working directory of a previous EnergyPlus simulation

Clean working directory of an EnergyPlus simulation by deleting all input and output files of previous simulation.

clean_wd(path)

Arguments

path
An .idf or .imf file path.

Details

clean_wd() imitates the same process that EnergyPlus does whenever a new simulation is getting to start. It deletes all related output files that have the same name prefix as the input path. The input model itself and any weather file are not deleted. clean_wd() is called internally when running EnergyPlus models using run_idf() and run_multi().

Author

Hongyuan Jia

Examples

# \dontrun{
# run a test simulation
idf_path <- system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr")
epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData",
     "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
)
dir <- file.path(tempdir(), "test")
run_idf(idf_path, epw_path, output_dir = dir, echo = FALSE)

#> $idf
#> [1] "/tmp/Rtmpa73RLC/temp_libpath50a7953f9d7/eplusr/extdata/1ZoneUncontrolled.idf"
#> 
#> $epw
#> [1] "/usr/local/EnergyPlus-8-8-0/WeatherData/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
#> 
#> $version
#> [1] "8.8.0"
#> 
#> $exit_status
#> [1] 0
#> 
#> $start_time
#> [1] "2020-11-11 02:04:44 UTC"
#> 
#> $end_time
#> [1] "2020-11-11 02:04:49 UTC"
#> 
#> $output_dir
#> [1] "/tmp/RtmpbSS17E/test"
#> 
#> $energyplus
#> [1] "/usr/local/EnergyPlus-8-8-0/energyplus-8.8.0"
#> 
#> $stdout
#>   [1] "EnergyPlus Starting"                                                                     
#>   [2] "EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2020.11.11 02:04"                              
#>   [3] "Processing Data Dictionary"                                                              
#>   [4] "Processing Input File"                                                                   
#>   [5] "Initializing Simulation"                                                                 
#>   [6] "Reporting Surfaces"                                                                      
#>   [7] "Beginning Primary Simulation"                                                            
#>   [8] "Initializing New Environment Parameters"                                                 
#>   [9] "Warming up {1}"                                                                          
#>  [10] "Warming up {2}"                                                                          
#>  [11] "Warming up {3}"                                                                          
#>  [12] "Warming up {4}"                                                                          
#>  [13] "Warming up {5}"                                                                          
#>  [14] "Warming up {6}"                                                                          
#>  [15] "Warming up {7}"                                                                          
#>  [16] "Warming up {8}"                                                                          
#>  [17] "Warming up {9}"                                                                          
#>  [18] "Warming up {10}"                                                                         
#>  [19] "Warming up {11}"                                                                         
#>  [20] "Warming up {12}"                                                                         
#>  [21] "Warming up {13}"                                                                         
#>  [22] "Warming up {14}"                                                                         
#>  [23] "Warming up {15}"                                                                         
#>  [24] "Warming up {16}"                                                                         
#>  [25] "Warming up {17}"                                                                         
#>  [26] "Warming up {18}"                                                                         
#>  [27] "Warming up {19}"                                                                         
#>  [28] "Warming up {20}"                                                                         
#>  [29] "Warming up {21}"                                                                         
#>  [30] "Warming up {22}"                                                                         
#>  [31] "Starting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB"    
#>  [32] "Initializing New Environment Parameters"                                                 
#>  [33] "Warming up {1}"                                                                          
#>  [34] "Warming up {2}"                                                                          
#>  [35] "Warming up {3}"                                                                          
#>  [36] "Warming up {4}"                                                                          
#>  [37] "Warming up {5}"                                                                          
#>  [38] "Warming up {6}"                                                                          
#>  [39] "Warming up {7}"                                                                          
#>  [40] "Warming up {8}"                                                                          
#>  [41] "Warming up {9}"                                                                          
#>  [42] "Warming up {10}"                                                                         
#>  [43] "Warming up {11}"                                                                         
#>  [44] "Warming up {12}"                                                                         
#>  [45] "Warming up {13}"                                                                         
#>  [46] "Warming up {14}"                                                                         
#>  [47] "Warming up {15}"                                                                         
#>  [48] "Warming up {16}"                                                                         
#>  [49] "Starting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB"
#>  [50] "Initializing New Environment Parameters"                                                 
#>  [51] "Warming up {1}"                                                                          
#>  [52] "Warming up {2}"                                                                          
#>  [53] "Warming up {3}"                                                                          
#>  [54] "Warming up {4}"                                                                          
#>  [55] "Warming up {5}"                                                                          
#>  [56] "Warming up {6}"                                                                          
#>  [57] "Warming up {7}"                                                                          
#>  [58] "Warming up {8}"                                                                          
#>  [59] "Warming up {9}"                                                                          
#>  [60] "Warming up {10}"                                                                         
#>  [61] "Warming up {11}"                                                                         
#>  [62] "Warming up {12}"                                                                         
#>  [63] "Warming up {13}"                                                                         
#>  [64] "Warming up {14}"                                                                         
#>  [65] "Warming up {15}"                                                                         
#>  [66] "Warming up {16}"                                                                         
#>  [67] "Warming up {17}"                                                                         
#>  [68] "Warming up {18}"                                                                         
#>  [69] "Starting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"          
#>  [70] "Updating Shadowing Calculations, Start Date=01/21"                                       
#>  [71] "Continuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [72] "Updating Shadowing Calculations, Start Date=02/10"                                       
#>  [73] "Continuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [74] "Updating Shadowing Calculations, Start Date=03/02"                                       
#>  [75] "Continuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [76] "Updating Shadowing Calculations, Start Date=03/22"                                       
#>  [77] "Continuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [78] "Updating Shadowing Calculations, Start Date=04/11"                                       
#>  [79] "Continuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [80] "Updating Shadowing Calculations, Start Date=05/01"                                       
#>  [81] "Continuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [82] "Updating Shadowing Calculations, Start Date=05/21"                                       
#>  [83] "Continuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [84] "Updating Shadowing Calculations, Start Date=06/10"                                       
#>  [85] "Continuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [86] "Updating Shadowing Calculations, Start Date=06/30"                                       
#>  [87] "Continuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [88] "Updating Shadowing Calculations, Start Date=07/20"                                       
#>  [89] "Continuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [90] "Updating Shadowing Calculations, Start Date=08/09"                                       
#>  [91] "Continuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [92] "Updating Shadowing Calculations, Start Date=08/29"                                       
#>  [93] "Continuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [94] "Updating Shadowing Calculations, Start Date=09/18"                                       
#>  [95] "Continuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [96] "Updating Shadowing Calculations, Start Date=10/08"                                       
#>  [97] "Continuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#>  [98] "Updating Shadowing Calculations, Start Date=10/28"                                       
#>  [99] "Continuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#> [100] "Updating Shadowing Calculations, Start Date=11/17"                                       
#> [101] "Continuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#> [102] "Updating Shadowing Calculations, Start Date=12/07"                                       
#> [103] "Continuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#> [104] "Updating Shadowing Calculations, Start Date=12/27"                                       
#> [105] "Continuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940"        
#> [106] "Writing tabular output file results using comma format."                                 
#> [107] "Writing tabular output file results using tab format."                                   
#> [108] "Writing tabular output file results using text format."                                  
#> [109] "Writing tabular output file results using HTML format."                                  
#> [110] "Writing tabular output file results using XML format."                                   
#> [111] "Writing final SQL reports"                                                               
#> [112] " ReadVarsESO program starting."                                                          
#> [113] " ReadVars Run Time=00hr 00min  0.38sec"                                                  
#> [114] " ReadVarsESO program completed successfully."                                            
#> [115] " ReadVarsESO program starting."                                                          
#> [116] " ReadVars Run Time=00hr 00min  0.17sec"                                                  
#> [117] " ReadVarsESO program completed successfully."                                            
#> [118] "EnergyPlus Run Time=00hr 00min  5.77sec"                                                 
#> 
#> $stderr
#> NULL
#> 
#> $process
#> PROCESS 'energyplus-8.8.0', finished.
#> 

list.files(dir)

#>  [1] "1ZoneUncontrolled.audit"                     
#>  [2] "1ZoneUncontrolled.bnd"                       
#>  [3] "1ZoneUncontrolled.csv"                       
#>  [4] "1ZoneUncontrolled.dxf"                       
#>  [5] "1ZoneUncontrolled.eio"                       
#>  [6] "1ZoneUncontrolled.end"                       
#>  [7] "1ZoneUncontrolled.err"                       
#>  [8] "1ZoneUncontrolled.eso"                       
#>  [9] "1ZoneUncontrolled.idf"                       
#> [10] "1ZoneUncontrolled.mdd"                       
#> [11] "1ZoneUncontrolled.mtd"                       
#> [12] "1ZoneUncontrolled.mtr"                       
#> [13] "1ZoneUncontrolled.rdd"                       
#> [14] "1ZoneUncontrolled.rvaudit"                   
#> [15] "1ZoneUncontrolled.shd"                       
#> [16] "1ZoneUncontrolledMeter.csv"                  
#> [17] "1ZoneUncontrolledSqlite.err"                 
#> [18] "1ZoneUncontrolledTable.csv"                  
#> [19] "1ZoneUncontrolledTable.htm"                  
#> [20] "1ZoneUncontrolledTable.tab"                  
#> [21] "1ZoneUncontrolledTable.txt"                  
#> [22] "1ZoneUncontrolledTable.xml"                  
#> [23] "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

# remove all output files
clean_wd(file.path(dir, basename(idf_path)))

list.files(dir)

#> [1] "1ZoneUncontrolled.idf"                       
#> [2] "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

# }