Run EnergyPlus and its various processors

Run EnergyPlus and its various processors

EPMacro(
  model,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

ExpandObjects(
  model,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL,
  idd = NULL
)

Basement(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL,
  idd = NULL
)

Slab(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL,
  idd = NULL
)

EnergyPlus(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  output_suffix = c("C", "L", "D"),
  wait = TRUE,
  echo = TRUE,
  annual = FALSE,
  design_day = FALSE,
  idd = NULL,
  eplus = NULL
)

convertESOMTR(
  eso,
  output_dir = NULL,
  output_prefix = NULL,
  rules = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

ReadVarsESO(
  eso,
  output_dir = NULL,
  output_prefix = NULL,
  output_suffix = c("C", "L", "D"),
  max_col = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

HVAC_Diagram(
  bnd,
  output_dir = NULL,
  output_prefix = NULL,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

energyplus(
  model,
  weather,
  output_dir = NULL,
  output_prefix = NULL,
  output_suffix = c("C", "L", "D"),
  epmacro = TRUE,
  expand_obj = TRUE,
  annual = FALSE,
  design_day = FALSE,
  eso_to_ip = FALSE,
  readvars = TRUE,
  echo = TRUE,
  wait = TRUE,
  idd = NULL,
  eplus = NULL,
  resources = NULL
)

Arguments

model

[character(1)]
A path of an EnergyPlus IDF or IMF file.

output_dir

[character(1) or NULL]
Output directory of EnergyPlus simulation outputs. If NULL, the directory where the input model locates is used. Default: NULL.

output_prefix

[character(1) or NULL]
Prefix for EnergyPlus output file names. If NULL, the input model file name is used. Default: NULL.

wait

[logical(1)]
If FALSE, simulation is run in the background and a processx::process object is returned. Extra steps are needed to collect the results after the process completes.

echo

[logical(1)]
Wheter to show standard output and error from EnergyPlus and its pre- and post- processors. Default: TRUE.

eplus

[character(1) or NULL]
An EnergyPlus version or a path of EnergyPlus installation directory. If NULL, the version of EnergyPlus to use is determined by the version of input model. Default: NULL.

idd

[character(1) or NULL]
The full path of EnergyPlus IDD (Input Data Dictionary). If NULL, Energy+.idd file in EnergyPlus installation directory is used. Default: NULL.

weather

[character(1) or NULL]
A path of an EnergyPlus weather (EPW) file. If NULL, design-day-only simulation is triggered, regardless of the design_day value.

output_suffix

[character(1)]
Suffix style for EnergyPlus output file names. Should be one of the followings:

C: Capital, e.g. eplusTable.csv. This is the default.

L: Legacy, e.g. eplustbl.csv.

D: Dash, e.g. eplus-table.csv.

annual

[logical(1)]
If TRUE, annual simulation is forced. Currently, only support EnergyPlus >= v8.3. Note that annual and design_day cannot both be TRUE. Default: FALSE.

design_day

[logical(1)]
If TRUE, design-day-only simulation is forced. Currently, only support EnergyPlus >= v8.3. Note that annual and design_day cannot both be TRUE. Default: FALSE.

eso

[character(1)]
A path of an EnergyPlus standard output (.eso) or EnergyPlus meter output (.mtr) file.

bnd

[character(1)]
A path of an EnergyPlus branch node details (.bnd) file.

epmacro

[logical(1)]
If TRUE, EPMacro processor is called perior to simulation. Only applicable if input file is an IMF file. Default: TRUE.

expand_obj

[logical(1)]
If TRUE, ExpandObjects processor is called perior to simulation. Should be TRUE if calling Basement or Slab preprocessors is desired. Default: TRUE.

eso_to_ip

[logical(1)]
If TRUE, convertESOMTR post-processor is called after simulation to convert the units of data in eso file from SI units to IP units. Default: FALSE.

readvars

[logical(1)]
If TRUE, ReadVarsESO post-processor is called after to simulation. Default: TRUE.

resources

[character() or NULL]
Any external file dependencies that EnergyPlus will use for simulation. If not NULL, files will be copied to the output directory. Default: NULL.

Value

Functions except for energyplus() return a list of two elements:

file: a named list of full paths of output files

run: a named list of outputs from the process.

energyplus() returns a list of 7 elements:

ver: EnergyPlus version used

energyplus: EnergyPlus installation directory

start_time: a POSIXct() giving the local time when the simulation starts

end_time: a POSIXct() giving the local time when the simulation ends

output_dir: full path of output directory of simulation outputs

file: a named list of relative paths of output files under output_dir

run: a data.table of each outputs from the all called processes

Details

EPMacro() calls the EnergyPlus EPMacro processor.

ExpandObjects() calls the EnergyPlus ExpandObjects processor.

Basement() calls the EnergyPlus Basement preprocessor.

Slab() calls the EnergyPlus Slab preprocessor.

EnergyPlus() calls EnergyPlus itself.

convertESOMTR() calls EnergyPlus convertESOMTR post-processor.

ReadVarsESO() calls EnergyPlus ReadVarsESO post-processor.

HVAC_Diagram() calls EnergyPlus HVAC-Diagram post-processor.

energyplus() is the one which correctly chains all the steps that call those pre- and post- processors to form a complete EnergyPlus simulation.

Note

energyplus() can only run in waiting mode.