eplusr: A Toolkit for Using EnergyPlus in R

A rich toolkit of using the whole building simulation program 'EnergyPlus'(<https://energyplus.net>), which enables programmatic navigation, modification of 'EnergyPlus' models and makes it less painful to do parametric simulations and analysis.

Details

eplusr provides a rich toolkit of using EnergyPlus directly in R, which enables programmatic navigation, modification of EnergyPlus models and makes it less painful to do parametric simulations and analysis.

With eplusr, you can do:

Read, parse and modify EnergyPlus Weather File (EPW).

Read and parse EnergyPlus IDF files.

Query on models, including classes, objects and fields

Directly add, modify, duplicate, and delete objects of IDF in R.

Automatically change referred fields when modifying objects.

Check any possible errors whenever modifications are made.

Save the changed models into standard formats in the same way as IDFEditor distributed along with EnergyPlus.

Run your models directly and collect the simulation output of EnergyPlus in R.

Run parametric analysis in parallel and collect results in one go.

See also