Get and Set eplusr options

Get and set eplusr options which affect the way in which eplusr computes and displays its results.

eplusr_option(...)

Arguments

...
Any options can be defined, using name = value. All available options are below. If no options are given, then all values of current options are returned.

Details

num_digits: Integer indicating the number of decimal places for numeric fields. Default: 8L

view_in_ip: Whether models should be presented in IP units. Default: FALSE

validate_level: The strictness level of validation during field value modification and model error checking. Possible value: "none", "draft" and "final". Default: "final". Detailed description:

For "none", none validation will be done;

For "draft", checking of invalid autosize, autocalculate, numeric, integer, and choice field values will be done;

For "final", besides above, checking of missing required objects, duplicated unique objects, object name conflicts, missing required fields and invalid field value reference will also be done.

verbose_info: Whether to show information messages. Default: TRUE.

save_format: The format to use when saving Idf objects to .idf files. Possible values: "asis", "sorted", "new_top" and "new_bottom". The later three have the same effect as Save Options settings "Sorted", "Original with New at Top" and "Original with New at Bottom" in IDF Editor, respectively. For "asis", the saving format will be set according to the header of IDF file. If no header found, "sorted" is used. "Default: "asis".

num_parallel: Maximum number of parallel simulations to run. Default: parallel::detectCores().