Format an Idf Object

Format an Idf object.

# S3 method for Idf
format(
  x,
  comment = TRUE,
  header = TRUE,
  format = eplusr_option("save_format"),
  leading = 4L,
  sep_at = 29L,
  ...
)

Arguments

x
An Idf object.

comment
If FALSE, all comments will not be included. Default: TRUE.

header
If FALSE, the header will not be included. Default: TRUE.

format
Specific format used when formatting. For details, please see $save(). Default: eplusr_option("save_format")

leading
Leading spaces added to each field. Default: 4L.

sep_at
The character width to separate value string and field string. Default: 29L which is the same as IDF Editor.

...
Further arguments passed to or from other methods.

Value

A single length string.

Author

Hongyuan Jia

Examples

# \dontrun{
idf_path <- system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr")
cat(format(read_idf(idf_path, use_idd(8.8, "auto")), leading = 0))

#> !-Generator eplusr
#> !-Option OriginalOrderTop
#> 
#> !-NOTE: All comments with '!-' are ignored by the IDFEditor and are generated automatically.
#> !-      Use '!' comments if they need to be retained when using the IDFEditor.
#> 
#> !1ZoneUncontrolled.idf
#> ! Basic file description:  Basic test for EnergyPlus.  Resistive Walls.  Regular (no ground contact) floor.
#> !                          Regular roof.  No Windows.
#> !
#> ! Highlights:              Very basic test to see that EnergyPlus "works".
#> !
#> !
#> ! Simulation Location/Run: Denver Centennial  Golden CO USA WMO=724666, 2 design days, 1 run period,
#> !                          Run Control executes two design days (see RUN PERIOD object)
#> !
#> ! Location:                Denver, CO
#> !
#> ! Design Days:             Denver Centennial  Golden CO USA Annual Heating 99%, MaxDB=-15.5C
#> !                          Denver Centennial  Golden CO USA Annual Cooling (DB=>MWB) 1%, MaxDB=32C MWB=15.5C
#> !
#> ! Run Period (Weather File): Full Annual Simulation, DENVER_STAPLETON_CO_USA_WMO_724690
#> !
#> ! Run Control:             No zone or system sizing, design day run control (no weather file simulation)
#> !
#> ! Building: Fictional 1 zone building with resistive walls.
#> !
#> !           The building is oriented due north.
#> !
#> ! Floor Area:        232.25 m2
#> ! Number of Stories: 1
#> !
#> ! Zone Description Details:
#> !
#> !       (0,15.24,0)                 (15.24,15.24,0)
#> !              _____________________________
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |                             |
#> !             |_____________________________|
#> !
#> !          (0,0,0)                      (15.24,0,0)
#> !
#> ! Internal gains description:     NA
#> !
#> ! Interzone Surfaces:             None
#> ! Internal Mass:                  None
#> ! People:                         None
#> ! Lights:                         None
#> ! Equipment:                      None
#> ! Windows:                        0
#> ! Detached Shading:               None
#> ! Daylight:                       None
#> ! Natural Ventilation:            None
#> ! Compact Schedules:              NA (Example of non-Compact Schedules)
#> ! Solar Distribution:             MinimalShadowing
#> !
#> ! HVAC:                           NA
#> !
#> ! Zonal Equipment:                NA
#> ! Central Air Handling Equipment: No
#> ! System Equipment Autosize:      No
#> ! Purchased Cooling:              No
#> ! Purchased Heating:              No
#> ! Purchased Chilled Water:        No
#> ! Purchased Hot Water:            No
#> ! Coils:                          None
#> ! Pumps:                          None
#> ! Boilers:                        None
#> ! Chillers:                       None
#> ! Towers:                         None
#> !
#> ! Results:
#> ! Standard Reports:               Variable Dictionary, Surfaces (dxf-wireframe), Meter File
#> ! Timestep or Hourly Variables:   Hourly and Daily
#> ! Time bins Report:               None
#> ! HTML Report:                    None
#> ! Environmental Emissions:        None
#> ! Utility Tariffs:                None
#> Version,
#> 8.8;                         !- Version Identifier
#> 
#> Timestep,
#> 4;                           !- Number of Timesteps per Hour
#> 
#> Building,
#> Simple One Zone (Wireframe DXF),  !- Name
#> 0,                           !- North Axis {deg}
#> Suburbs,                     !- Terrain
#> 0.04,                        !- Loads Convergence Tolerance Value
#> 0.004,                       !- Temperature Convergence Tolerance Value {deltaC}
#> MinimalShadowing,            !- Solar Distribution
#> 30,                          !- Maximum Number of Warmup Days
#> 6;                           !- Minimum Number of Warmup Days
#> 
#> HeatBalanceAlgorithm,
#> ConductionTransferFunction;  !- Algorithm
#> 
#> SurfaceConvectionAlgorithm:Inside,
#> TARP;                        !- Algorithm
#> 
#> SurfaceConvectionAlgorithm:Outside,
#> DOE-2;                       !- Algorithm
#> 
#> SimulationControl,
#> No,                          !- Do Zone Sizing Calculation
#> No,                          !- Do System Sizing Calculation
#> No,                          !- Do Plant Sizing Calculation
#> Yes,                         !- Run Simulation for Sizing Periods
#> Yes;                         !- Run Simulation for Weather File Run Periods
#> 
#> RunPeriod,
#> ,                            !- Name
#> 1,                           !- Begin Month
#> 1,                           !- Begin Day of Month
#> 12,                          !- End Month
#> 31,                          !- End Day of Month
#> Tuesday,                     !- Day of Week for Start Day
#> Yes,                         !- Use Weather File Holidays and Special Days
#> Yes,                         !- Use Weather File Daylight Saving Period
#> No,                          !- Apply Weekend Holiday Rule
#> Yes,                         !- Use Weather File Rain Indicators
#> Yes;                         !- Use Weather File Snow Indicators
#> 
#> Site:Location,
#> Denver Centennial  Golden   N_CO_USA Design_Conditions,  !- Name
#> 39.74,                       !- Latitude {deg}
#> -105.18,                     !- Longitude {deg}
#> -7,                          !- Time Zone {hr}
#> 1829;                        !- Elevation {m}
#> 
#> !  WMO=724666 Time Zone=NAM: (GMT-07:00) Mountain Time (US & Canada)
#> !  Data Source=ASHRAE 2009 Annual Design Conditions
#> ! Denver Centennial  Golden   N_CO_USA Annual Heating Design Conditions Wind Speed=3m/s Wind Dir=340
#> ! Coldest Month=DEC
#> ! Denver Centennial  Golden   N_CO_USA Annual Heating 99%, MaxDB=-15.5C
#> SizingPeriod:DesignDay,
#> Denver Centennial  Golden   N Ann Htg 99% Condns DB,  !- Name
#> 12,                          !- Month
#> 21,                          !- Day of Month
#> WinterDesignDay,             !- Day Type
#> -15.5,                       !- Maximum Dry-Bulb Temperature {C}
#> 0,                           !- Daily Dry-Bulb Temperature Range {deltaC}
#> ,                            !- Dry-Bulb Temperature Range Modifier Type
#> ,                            !- Dry-Bulb Temperature Range Modifier Day Schedule Name
#> Wetbulb,                     !- Humidity Condition Type
#> -15.5,                       !- Wetbulb or DewPoint at Maximum Dry-Bulb {C}
#> ,                            !- Humidity Condition Day Schedule Name
#> ,                            !- Humidity Ratio at Maximum Dry-Bulb {kgWater/kgDryAir}
#> ,                            !- Enthalpy at Maximum Dry-Bulb {J/kg}
#> ,                            !- Daily Wet-Bulb Temperature Range {deltaC}
#> 81198,                       !- Barometric Pressure {Pa}
#> 3,                           !- Wind Speed {m/s}
#> 340,                         !- Wind Direction {deg}
#> No,                          !- Rain Indicator
#> No,                          !- Snow Indicator
#> No,                          !- Daylight Saving Time Indicator
#> ASHRAEClearSky,              !- Solar Model Indicator
#> ,                            !- Beam Solar Day Schedule Name
#> ,                            !- Diffuse Solar Day Schedule Name
#> ,                            !- ASHRAE Clear Sky Optical Depth for Beam Irradiance (taub) {dimensionless}
#> ,                            !- ASHRAE Clear Sky Optical Depth for Diffuse Irradiance (taud) {dimensionless}
#> 0;                           !- Sky Clearness
#> 
#> ! Denver Centennial  Golden   N Annual Cooling Design Conditions Wind Speed=4.9m/s Wind Dir=0
#> ! Hottest Month=JUL
#> ! Denver Centennial  Golden   N_CO_USA Annual Cooling (DB=>MWB) 1%, MaxDB=32C MWB=15.5C
#> SizingPeriod:DesignDay,
#> Denver Centennial  Golden   N Ann Clg 1% Condns DB=>MWB,  !- Name
#> 7,                           !- Month
#> 21,                          !- Day of Month
#> SummerDesignDay,             !- Day Type
#> 32,                          !- Maximum Dry-Bulb Temperature {C}
#> 15.2,                        !- Daily Dry-Bulb Temperature Range {deltaC}
#> ,                            !- Dry-Bulb Temperature Range Modifier Type
#> ,                            !- Dry-Bulb Temperature Range Modifier Day Schedule Name
#> Wetbulb,                     !- Humidity Condition Type
#> 15.5,                        !- Wetbulb or DewPoint at Maximum Dry-Bulb {C}
#> ,                            !- Humidity Condition Day Schedule Name
#> ,                            !- Humidity Ratio at Maximum Dry-Bulb {kgWater/kgDryAir}
#> ,                            !- Enthalpy at Maximum Dry-Bulb {J/kg}
#> ,                            !- Daily Wet-Bulb Temperature Range {deltaC}
#> 81198,                       !- Barometric Pressure {Pa}
#> 4.9,                         !- Wind Speed {m/s}
#> 0,                           !- Wind Direction {deg}
#> No,                          !- Rain Indicator
#> No,                          !- Snow Indicator
#> No,                          !- Daylight Saving Time Indicator
#> ASHRAEClearSky,              !- Solar Model Indicator
#> ,                            !- Beam Solar Day Schedule Name
#> ,                            !- Diffuse Solar Day Schedule Name
#> ,                            !- ASHRAE Clear Sky Optical Depth for Beam Irradiance (taub) {dimensionless}
#> ,                            !- ASHRAE Clear Sky Optical Depth for Diffuse Irradiance (taud) {dimensionless}
#> 1;                           !- Sky Clearness
#> 
#> Material:NoMass,
#> R13LAYER,                    !- Name
#> Rough,                       !- Roughness
#> 2.290965,                    !- Thermal Resistance {m2-K/W}
#> 0.9,                         !- Thermal Absorptance
#> 0.75,                        !- Solar Absorptance
#> 0.75;                        !- Visible Absorptance
#> 
#> Material:NoMass,
#> R31LAYER,                    !- Name
#> Rough,                       !- Roughness
#> 5.456,                       !- Thermal Resistance {m2-K/W}
#> 0.9,                         !- Thermal Absorptance
#> 0.75,                        !- Solar Absorptance
#> 0.75;                        !- Visible Absorptance
#> 
#> Material,
#> C5 - 4 IN HW CONCRETE,       !- Name
#> MediumRough,                 !- Roughness
#> 0.1014984,                   !- Thickness {m}
#> 1.729577,                    !- Conductivity {W/m-K}
#> 2242.585,                    !- Density {kg/m3}
#> 836.8,                       !- Specific Heat {J/kg-K}
#> 0.9,                         !- Thermal Absorptance
#> 0.65,                        !- Solar Absorptance
#> 0.65;                        !- Visible Absorptance
#> 
#> Construction,
#> R13WALL,                     !- Name
#> R13LAYER;                    !- Outside Layer
#> 
#> Construction,
#> FLOOR,                       !- Name
#> C5 - 4 IN HW CONCRETE;       !- Outside Layer
#> 
#> Construction,
#> ROOF31,                      !- Name
#> R31LAYER;                    !- Outside Layer
#> 
#> Zone,
#> ZONE ONE,                    !- Name
#> 0,                           !- Direction of Relative North {deg}
#> 0,                           !- X Origin {m}
#> 0,                           !- Y Origin {m}
#> 0,                           !- Z Origin {m}
#> 1,                           !- Type
#> 1,                           !- Multiplier
#> autocalculate,               !- Ceiling Height {m}
#> autocalculate;               !- Volume {m3}
#> 
#> ScheduleTypeLimits,
#> Fraction,                    !- Name
#> 0,                           !- Lower Limit Value
#> 1,                           !- Upper Limit Value
#> CONTINUOUS;                  !- Numeric Type
#> 
#> GlobalGeometryRules,
#> UpperLeftCorner,             !- Starting Vertex Position
#> CounterClockWise,            !- Vertex Entry Direction
#> World;                       !- Coordinate System
#> 
#> BuildingSurface:Detailed,
#> Zn001:Wall001,               !- Name
#> Wall,                        !- Surface Type
#> R13WALL,                     !- Construction Name
#> ZONE ONE,                    !- Zone Name
#> Outdoors,                    !- Outside Boundary Condition
#> ,                            !- Outside Boundary Condition Object
#> SunExposed,                  !- Sun Exposure
#> WindExposed,                 !- Wind Exposure
#> 0.5,                         !- View Factor to Ground
#> 4,                           !- Number of Vertices
#> 0,                           !- Vertex 1 X-coordinate {m}
#> 0,                           !- Vertex 1 Y-coordinate {m}
#> 4.572,                       !- Vertex 1 Z-coordinate {m}
#> 0,                           !- Vertex 2 X-coordinate {m}
#> 0,                           !- Vertex 2 Y-coordinate {m}
#> 0,                           !- Vertex 2 Z-coordinate {m}
#> 15.24,                       !- Vertex 3 X-coordinate {m}
#> 0,                           !- Vertex 3 Y-coordinate {m}
#> 0,                           !- Vertex 3 Z-coordinate {m}
#> 15.24,                       !- Vertex 4 X-coordinate {m}
#> 0,                           !- Vertex 4 Y-coordinate {m}
#> 4.572;                       !- Vertex 4 Z-coordinate {m}
#> 
#> BuildingSurface:Detailed,
#> Zn001:Wall002,               !- Name
#> Wall,                        !- Surface Type
#> R13WALL,                     !- Construction Name
#> ZONE ONE,                    !- Zone Name
#> Outdoors,                    !- Outside Boundary Condition
#> ,                            !- Outside Boundary Condition Object
#> SunExposed,                  !- Sun Exposure
#> WindExposed,                 !- Wind Exposure
#> 0.5,                         !- View Factor to Ground
#> 4,                           !- Number of Vertices
#> 15.24,                       !- Vertex 1 X-coordinate {m}
#> 0,                           !- Vertex 1 Y-coordinate {m}
#> 4.572,                       !- Vertex 1 Z-coordinate {m}
#> 15.24,                       !- Vertex 2 X-coordinate {m}
#> 0,                           !- Vertex 2 Y-coordinate {m}
#> 0,                           !- Vertex 2 Z-coordinate {m}
#> 15.24,                       !- Vertex 3 X-coordinate {m}
#> 15.24,                       !- Vertex 3 Y-coordinate {m}
#> 0,                           !- Vertex 3 Z-coordinate {m}
#> 15.24,                       !- Vertex 4 X-coordinate {m}
#> 15.24,                       !- Vertex 4 Y-coordinate {m}
#> 4.572;                       !- Vertex 4 Z-coordinate {m}
#> 
#> BuildingSurface:Detailed,
#> Zn001:Wall003,               !- Name
#> Wall,                        !- Surface Type
#> R13WALL,                     !- Construction Name
#> ZONE ONE,                    !- Zone Name
#> Outdoors,                    !- Outside Boundary Condition
#> ,                            !- Outside Boundary Condition Object
#> SunExposed,                  !- Sun Exposure
#> WindExposed,                 !- Wind Exposure
#> 0.5,                         !- View Factor to Ground
#> 4,                           !- Number of Vertices
#> 15.24,                       !- Vertex 1 X-coordinate {m}
#> 15.24,                       !- Vertex 1 Y-coordinate {m}
#> 4.572,                       !- Vertex 1 Z-coordinate {m}
#> 15.24,                       !- Vertex 2 X-coordinate {m}
#> 15.24,                       !- Vertex 2 Y-coordinate {m}
#> 0,                           !- Vertex 2 Z-coordinate {m}
#> 0,                           !- Vertex 3 X-coordinate {m}
#> 15.24,                       !- Vertex 3 Y-coordinate {m}
#> 0,                           !- Vertex 3 Z-coordinate {m}
#> 0,                           !- Vertex 4 X-coordinate {m}
#> 15.24,                       !- Vertex 4 Y-coordinate {m}
#> 4.572;                       !- Vertex 4 Z-coordinate {m}
#> 
#> BuildingSurface:Detailed,
#> Zn001:Wall004,               !- Name
#> Wall,                        !- Surface Type
#> R13WALL,                     !- Construction Name
#> ZONE ONE,                    !- Zone Name
#> Outdoors,                    !- Outside Boundary Condition
#> ,                            !- Outside Boundary Condition Object
#> SunExposed,                  !- Sun Exposure
#> WindExposed,                 !- Wind Exposure
#> 0.5,                         !- View Factor to Ground
#> 4,                           !- Number of Vertices
#> 0,                           !- Vertex 1 X-coordinate {m}
#> 15.24,                       !- Vertex 1 Y-coordinate {m}
#> 4.572,                       !- Vertex 1 Z-coordinate {m}
#> 0,                           !- Vertex 2 X-coordinate {m}
#> 15.24,                       !- Vertex 2 Y-coordinate {m}
#> 0,                           !- Vertex 2 Z-coordinate {m}
#> 0,                           !- Vertex 3 X-coordinate {m}
#> 0,                           !- Vertex 3 Y-coordinate {m}
#> 0,                           !- Vertex 3 Z-coordinate {m}
#> 0,                           !- Vertex 4 X-coordinate {m}
#> 0,                           !- Vertex 4 Y-coordinate {m}
#> 4.572;                       !- Vertex 4 Z-coordinate {m}
#> 
#> BuildingSurface:Detailed,
#> Zn001:Flr001,                !- Name
#> Floor,                       !- Surface Type
#> FLOOR,                       !- Construction Name
#> ZONE ONE,                    !- Zone Name
#> Adiabatic,                   !- Outside Boundary Condition
#> ,                            !- Outside Boundary Condition Object
#> NoSun,                       !- Sun Exposure
#> NoWind,                      !- Wind Exposure
#> 1,                           !- View Factor to Ground
#> 4,                           !- Number of Vertices
#> 15.24,                       !- Vertex 1 X-coordinate {m}
#> 0,                           !- Vertex 1 Y-coordinate {m}
#> 0,                           !- Vertex 1 Z-coordinate {m}
#> 0,                           !- Vertex 2 X-coordinate {m}
#> 0,                           !- Vertex 2 Y-coordinate {m}
#> 0,                           !- Vertex 2 Z-coordinate {m}
#> 0,                           !- Vertex 3 X-coordinate {m}
#> 15.24,                       !- Vertex 3 Y-coordinate {m}
#> 0,                           !- Vertex 3 Z-coordinate {m}
#> 15.24,                       !- Vertex 4 X-coordinate {m}
#> 15.24,                       !- Vertex 4 Y-coordinate {m}
#> 0;                           !- Vertex 4 Z-coordinate {m}
#> 
#> BuildingSurface:Detailed,
#> Zn001:Roof001,               !- Name
#> Roof,                        !- Surface Type
#> ROOF31,                      !- Construction Name
#> ZONE ONE,                    !- Zone Name
#> Outdoors,                    !- Outside Boundary Condition
#> ,                            !- Outside Boundary Condition Object
#> SunExposed,                  !- Sun Exposure
#> WindExposed,                 !- Wind Exposure
#> 0,                           !- View Factor to Ground
#> 4,                           !- Number of Vertices
#> 0,                           !- Vertex 1 X-coordinate {m}
#> 15.24,                       !- Vertex 1 Y-coordinate {m}
#> 4.572,                       !- Vertex 1 Z-coordinate {m}
#> 0,                           !- Vertex 2 X-coordinate {m}
#> 0,                           !- Vertex 2 Y-coordinate {m}
#> 4.572,                       !- Vertex 2 Z-coordinate {m}
#> 15.24,                       !- Vertex 3 X-coordinate {m}
#> 0,                           !- Vertex 3 Y-coordinate {m}
#> 4.572,                       !- Vertex 3 Z-coordinate {m}
#> 15.24,                       !- Vertex 4 X-coordinate {m}
#> 15.24,                       !- Vertex 4 Y-coordinate {m}
#> 4.572;                       !- Vertex 4 Z-coordinate {m}
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Site Outdoor Air Drybulb Temperature,  !- Variable Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Site Daylight Saving Time Status,  !- Variable Name
#> daily;                       !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Site Day Type Index,         !- Variable Name
#> daily;                       !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Zone Mean Air Temperature,   !- Variable Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Zone Total Internal Latent Gain Energy,  !- Variable Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Zone Mean Radiant Temperature,  !- Variable Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Zone Air Heat Balance Surface Convection Rate,  !- Variable Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Zone Air Heat Balance Air Energy Storage Rate,  !- Variable Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Surface Inside Face Temperature,  !- Variable Name
#> daily;                       !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Surface Outside Face Temperature,  !- Variable Name
#> daily;                       !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Surface Inside Face Convection Heat Transfer Coefficient,  !- Variable Name
#> daily;                       !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Surface Outside Face Convection Heat Transfer Coefficient,  !- Variable Name
#> daily;                       !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Other Equipment Total Heating Energy,  !- Variable Name
#> monthly;                     !- Reporting Frequency
#> 
#> Output:Variable,
#> *,                           !- Key Value
#> Zone Other Equipment Total Heating Energy,  !- Variable Name
#> monthly;                     !- Reporting Frequency
#> 
#> Output:VariableDictionary,
#> IDF;                         !- Key Field
#> 
#> Output:Surfaces:Drawing,
#> dxf:wireframe;               !- Report Type
#> 
#> Output:Constructions,
#> Constructions;               !- Details Type 1
#> 
#> Output:Meter:MeterFileOnly,
#> ExteriorLights:Electricity,  !- Key Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Meter:MeterFileOnly,
#> EnergyTransfer:Building,     !- Key Name
#> hourly;                      !- Reporting Frequency
#> 
#> Output:Meter:MeterFileOnly,
#> EnergyTransfer:Facility,     !- Key Name
#> hourly;                      !- Reporting Frequency
#> 
#> OutputControl:Table:Style,
#> ALL;                         !- Column Separator
#> 
#> Output:Table:SummaryReports,
#> AllSummary;                  !- Report 1 Name
#> 
#> Exterior:Lights,
#> ExtLights,                   !- Name
#> AlwaysOn,                    !- Schedule Name
#> 5250,                        !- Design Level {W}
#> AstronomicalClock,           !- Control Option
#> Grounds Lights;              !- End-Use Subcategory
#> 
#> ScheduleTypeLimits,
#> On/Off,                      !- Name
#> 0,                           !- Lower Limit Value
#> 1,                           !- Upper Limit Value
#> DISCRETE;                    !- Numeric Type
#> 
#> OtherEquipment,
#> Test 352a,                   !- Name
#> None,                        !- Fuel Type
#> ZONE ONE,                    !- Zone or ZoneList Name
#> AlwaysOn,                    !- Schedule Name
#> EquipmentLevel,              !- Design Level Calculation Method
#> 352,                         !- Design Level {W}
#> ,                            !- Power per Zone Floor Area {W/m2}
#> ,                            !- Power per Person {W/person}
#> 0,                           !- Fraction Latent
#> 0,                           !- Fraction Radiant
#> 0;                           !- Fraction Lost
#> 
#> OtherEquipment,
#> Test 352 minus,              !- Name
#> None,                        !- Fuel Type
#> ZONE ONE,                    !- Zone or ZoneList Name
#> AlwaysOn,                    !- Schedule Name
#> EquipmentLevel,              !- Design Level Calculation Method
#> -352,                        !- Design Level {W}
#> ,                            !- Power per Zone Floor Area {W/m2}
#> ,                            !- Power per Person {W/person}
#> 0,                           !- Fraction Latent
#> 0,                           !- Fraction Radiant
#> 0;                           !- Fraction Lost
#> 
#> Schedule:Constant,
#> AlwaysOn,                    !- Name
#> On/Off,                      !- Schedule Type Limits Name
#> 1;                           !- Hourly Value

# }