Extract object and value reference relations

Extract object and value reference relations

get_idf_relation(
  idd_env,
  idf_env,
  object_id = NULL,
  value_id = NULL,
  direction = c("ref_to", "ref_by"),
  depth = 0L,
  name = FALSE,
  object = NULL,
  class = NULL,
  group = NULL,
  keep_all = FALSE,
  class_ref = c("both", "none", "all"),
  match_all = FALSE
)

Arguments

idd_env
An environment or list contains IDD tables including class, field, and reference.

idf_env
An environment or list contains IDF tables including object, value, and reference.

object_id
An integer vector of valid object IDs. If NULL, all object IDs in current IDF will be used.

value_id
An integer vector of valid value IDs. If NULL, all value IDs in current IDF will be used.

direction
Reference relation direction. Should be one of "ref_to" and "ref_by". Default: "ref_to".

depth
Recursive reference relation depth. NULL means infinite. Default: 0L.

name
If TRUE, all class, object, field value ID and name columns will be added and a IdfRelationTo or IdfRelationBy object is returned with customized printing method. Default: FALSE.

object
An integer vector of valid object IDs or a character vector of valid object names to specify the targeting relation objects. Default: NULL.

class
An integer vector of valid class indexes or a character vector of valid class names to specify the targeting relation classes. Default: NULL.

group
A character vector of valid group names to specify the targeting relation groups. Default: NULL.

keep_all
If TRUE, all input ID are kept. Otherwise, only input IDs that have relations are kept. Default: FALSE.

class_ref
Specify how to handle class-name-references. There are 3 options in total, i.e. "none", "both" and "all", with "both" being the default. * "none": just ignore class-name-references. * "both": only include class-name-references if this object also reference field values of the same one. This is the default option. * "all": include all class-name-references. This is the most aggressive option.

match_all
If TRUE, relation search will continue even though one relation has been found. If FALSE, searching is stopped whenever one relation is found in specified classes/groups. Default: FALSE.

Value

A data.table.