Get value data

Get value data

get_idf_value(
  idd_env,
  idf_env,
  class = NULL,
  object = NULL,
  field = NULL,
  property = NULL,
  underscore = FALSE,
  ignore_case = FALSE,
  align = FALSE,
  complete = FALSE,
  all = FALSE
)

Arguments

idd_env

An environment or list contains IDD tables including class, field, and reference.

idf_env

An environment or list contains IDF tables including object, value, and reference.

class

An integer vector of valid class indexes or a character vector of valid class names. Default: NULL.

object

An integer vector of valid object IDs or a character vector of valid object names. Default: NULL.

field

An integer vector of valid field indexes or a character vector of valid field names (can be in in underscore style). class and field should have the same length.

property

A character vector of column names in field table to return.

underscore

If TRUE, input class name will be converted into underscore style name first and column class_name_us will be used for matching. Default: FALSE.

ignore_case

If TRUE, input object name will be converted into lower case and column object_name_lower will be used for matching. converted into underscore style name first and column class_name_us and field_name_us will be used for matching. Default: FALSE.

align

If TRUE, all objects in the same class will have the same field number. The number of fields is the same as the object that have the most fields among objects specified. Default: FALSE.

complete

If TRUE, at least fields till the current whole extensible group will be returned. A new column named "matched_rleid" will be created (when property is NULL) indicating if given field has been matched or not.

all

If TRUE, all available fields defined in IDD for the class that objects belong to will be returned. Default: FALSE.

Value

A data.table containing specified columns.