Format object information string

Format object information string

get_object_info(
  dt_object,
  component = c("id", "name", "class"),
  by_class = FALSE,
  numbered = TRUE,
  collapse = NULL,
  prefix = NULL,
  name_prefix = TRUE
)

Arguments

dt_object
A data.table::data.table() of object data

component
A character vector specifying what information to be formatted. Should be a subset of "id", "name" and "class". Defaults are all of them.

by_class
If TRUE, multiple objects in the same class will be concatenated. Default: FALSE.

numbered
If TRUE, a index number will be prepended. If rleid column exists in dt_object, its values will be used as the index numbers.

collapse
A single string used to collapse the results into a single string. Default: NULL.

prefix
A character vector used to add at the beginning of object information. Default: NULL.

name_prefix
If TRUE, Default: TRUE.

Value

A character vector of the same length as the row number of input dt_object if collapse is NULL. Otherwise a single string.