Create and Run Parametric Analysis, and Collect Results

ParametricJob class provides a prototype of conducting parametric analysis of EnergyPlus simulations.

Details

Basically, it is a collection of multiple EplusJob objects.

Usage

param <- param_job(idf, epw)
param$apply_measure(measure, ..., .names = NULL)
param$run(dir = NULL, parallel_backend = future::multiprocess)
param$kill(which = NULL)
param$status(which = NULL)
param$errors(info = FALSE)
param$output_dir(which = NULL)
param$locate_output(which = NULL, suffix = ".err", strict = TRUE)
param$report_data_dict(which = NULL)
param$report_data(which = NULL, key_value = NULL, name = NULL, year = NULL, tz = "GMT", case = "auto")
param$tabular_data(which = NULL)
param$print()

Create

param <- param_job(idf, epw)

Arguments

idf: Path to EnergyPlus IDF or IMF file or an Idf object.

epw: Path to EnergyPlus EPW file or an Epw object.

Get Seed Model and Weather

param$seed()
param$weather()

$seed() will return the input Idf object.

$weather() will return the input Epw object.

Apply Design Alternatives

param$apply_measure(measure, ..., .names = NULL)

$apply_measure() allows to apply a measure to an Idf and creates parametric models for analysis. Basically, a measure is just a function that takes an Idf object and other arguments as input, and returns a modified Idf object as output. Use ... to supply different arguments to that measure. Under the hook, mapply() is used to create multiple Idfs according to the input values.

Arguments

measure: A function that takes an Idf and other arguments as input and returns an Idf object as output.

...: Other arguments passed to that measure.

.names: A character vector of the names of parametric Idfs. If NULL, the new Idfs will be named in format measure_name + number.

Run and Collect Results

param$run(dir = NULL, parallel_backend = future::multiprocess)
param$kill(which = NULL)
param$status(which = NULL)
param$errors(info = FALSE)
param$output_dir(which = NULL)
param$locate_output(which = NULL, suffix = ".err", strict = TRUE)
param$report_data_dict(which = NULL)
param$report_data(which = NULL, key_value = NULL, name = NULL, year = NULL, tz = "GMT", case = "auto")
param$tabular_data(which = NULL)

All those functions have the same meaning in EplusJob class, except that they only return the results of specified simulation.

Arguments

which: An integer vector of the indexes or a character vector or names of parametric simulations.

parallel_backend: Any acceptable input for future::plan().

All other arguments have the same meanings as in EplusJob class.