Read an EnergyPlus Report Data Dictionary File

read_rdd() takes a file path of EnergyPlus Report Data Dictionary (RDD) file, parses it and returns a RddFile object. read_mdd() takes a file path of EnergyPlus Meter Data Dictionary (MDD) file, parses it and returns a MddFile object.

read_rdd(path)

read_mdd(path)

Arguments

path
For read_rdd(), a file path of EnergyPlus EnergyPlus Report Data Dictionary file with an extension .rdd. For read_mdd(), a file path of EnergyPlus EnergyPlus Meter Data Dictionary file with an extension .mdd

Value

For read_rdd(), an RddFile object. For read_mdd(), a MddFile object.

Details

Basically, a RddFile and MddFile object is a data.table with 5 columns and 3 additional attributes:

5 Columns:

*index: Integer. Index of each variable.

reported_time_step: Character. Reported time step for the variables. Possible value: Zone and HVAC.

report_type: Character. Report types. Possible value: Average, Sum and Meter. Note that Meter is only for MDD file. All variables will have report_type being Meter.

variable: Character. Report variable names.

units: Character. Units of reported values. NA if report values do not have units.

3 Attributes:

eplus_version: A numeric_version object. The version of EnergyPlus used during the simulation.

eplus_build: A single string. The build tag of EnergyPlus used during the simulation.

datetime: A DateTime (POSIXct). The time when the simulation started.

Author

Hongyuan Jia

Examples

# \dontrun{
# run simulation and get the err file
idf_name <- "1ZoneUncontrolled.idf"
epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
idf_path <- file.path(eplus_config(8.8)$dir, "ExampleFiles", idf_name)
epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData", epw_name)
job <- eplus_job(idf_path, epw_path)

#> Adding an object in class 'Output:SQLite' and setting its 'Option Type' to 'SimpleAndTabular' in order to create SQLite output file.

job$run(dir = tempdir())

#> EnergyPlus Starting
#> EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2021.06.09 06:25
#> Processing Data Dictionary
#> Processing Input File
#> Initializing Simulation
#> Reporting Surfaces
#> Beginning Primary Simulation
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Warming up {19}
#> Warming up {20}
#> Warming up {21}
#> Warming up {22}
#> Starting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Starting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Starting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=01/21
#> Continuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=02/10
#> Continuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/02
#> Continuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/22
#> Continuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=04/11
#> Continuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/01
#> Continuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/21
#> Continuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/10
#> Continuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/30
#> Continuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=07/20
#> Continuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/09
#> Continuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/29
#> Continuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=09/18
#> Continuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/08
#> Continuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/28
#> Continuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=11/17
#> Continuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/07
#> Continuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/27
#> Continuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Writing tabular output file results using comma format.
#> Writing tabular output file results using tab format.
#> Writing tabular output file results using text format.
#> Writing tabular output file results using HTML format.
#> Writing tabular output file results using XML format.
#> Writing final SQL reports
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.25sec
#>  ReadVarsESO program completed successfully.
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.12sec
#>  ReadVarsESO program completed successfully.
#> EnergyPlus Run Time=00hr 00min  1.92sec
#> 
#> 

#> ── EnergPlus Simulation Job ────────────────────────────────────────────────────
#> * Model: '/tmp/RtmpA0bAlg/1ZoneUncontrolled.idf'
#> * Weather: '/home/runner/.local/EnergyPlus-8-8-0/WeatherData/USA_CA_San.Fra...
#> * EnergyPlus Version: '8.8.0'
#> * EnergyPlus Path: '/home/runner/.local/EnergyPlus-8-8-0'
#>  Simulation started at '2021-06-09 06:25:04' and completed successfully after 1.96 secs.

# read the err file
read_rdd(job$locate_output(".rdd"))

#> ══ EnergyPlus Report Data Dictionary File ══════════════════════════════════════
#>   * EnergyPlus version: 8.8.0 (7c3bbe4830)
#>   * Simulation started: 2021-06-09 06:25:00
#> 
#> ── Details ─────────────────────────────────────────────────────────────────────
#>      index reported_time_step report_type
#>   1:     1               Zone     Average
#>   2:     2               Zone     Average
#>   3:     3               Zone     Average
#>   4:     4               Zone     Average
#>   5:     5               Zone     Average
#>  ---                                     
#> 244:   244               Zone         Sum
#> 245:   245               Zone     Average
#> 246:   246               HVAC         Sum
#> 247:   247               HVAC         Sum
#> 248:   248               HVAC         Sum
#>                                                             variable
#>   1:                            Site Outdoor Air Drybulb Temperature
#>   2:                           Site Outdoor Air Dewpoint Temperature
#>   3:                            Site Outdoor Air Wetbulb Temperature
#>   4:                                 Site Outdoor Air Humidity Ratio
#>   5:                              Site Outdoor Air Relative Humidity
#>  ---                                                                
#> 244:           Facility Cooling Setpoint Not Met While Occupied Time
#> 245:                                                  Schedule Value
#> 246: Environmental Impact Total N2O Emissions Carbon Equivalent Mass
#> 247: Environmental Impact Total CH4 Emissions Carbon Equivalent Mass
#> 248: Environmental Impact Total CO2 Emissions Carbon Equivalent Mass
#>                 units
#>   1:                C
#>   2:                C
#>   3:                C
#>   4: kgWater/kgDryAir
#>   5:                %
#>  ---                 
#> 244:               hr
#> 245:             <NA>
#> 246:               kg
#> 247:               kg
#> 248:               kg

read_mdd(job$locate_output(".mdd"))

#> ══ EnergyPlus Meter Data Dictionary File ═══════════════════════════════════════
#>   * EnergyPlus version: 8.8.0 (7c3bbe4830)
#>   * Simulation started: 2021-06-09 06:25:00
#> 
#> ── Details ─────────────────────────────────────────────────────────────────────
#>     index reported_time_step report_type
#>  1:     1               Zone       Meter
#>  2:     2               Zone       Meter
#>  3:     3               Zone       Meter
#>  4:     4               Zone       Meter
#>  5:     5               Zone       Meter
#>  6:     6               Zone       Meter
#>  7:     7               Zone       Meter
#>  8:     8               Zone       Meter
#>  9:     9               Zone       Meter
#> 10:    10               Zone       Meter
#> 11:    11               Zone       Meter
#> 12:    12               Zone       Meter
#> 13:    13               Zone       Meter
#> 14:    14               Zone       Meter
#> 15:    15               Zone       Meter
#> 16:    16               Zone       Meter
#> 17:    17               Zone       Meter
#> 18:    18               Zone       Meter
#> 19:    19               Zone       Meter
#> 20:    20               Zone       Meter
#> 21:    21               Zone       Meter
#>     index reported_time_step report_type
#>                                        variable units
#>  1:                        Electricity:Facility     J
#>  2:                  ExteriorLights:Electricity     J
#>  3:   Grounds Lights:ExteriorLights:Electricity     J
#>  4:                     EnergyTransfer:Facility     J
#>  5:                     EnergyTransfer:Building     J
#>  6:                EnergyTransfer:Zone:ZONE ONE     J
#>  7:                      Heating:EnergyTransfer     J
#>  8:        Heating:EnergyTransfer:Zone:ZONE ONE     J
#>  9:                      Cooling:EnergyTransfer     J
#> 10:        Cooling:EnergyTransfer:Zone:ZONE ONE     J
#> 11:               ElectricityPurchased:Facility     J
#> 12:                  ElectricityPurchased:Plant     J
#> 13:           Cogeneration:ElectricityPurchased     J
#> 14:             ElectricitySurplusSold:Facility     J
#> 15:                ElectricitySurplusSold:Plant     J
#> 16:         Cogeneration:ElectricitySurplusSold     J
#> 17:                     ElectricityNet:Facility     J
#> 18:                        ElectricityNet:Plant     J
#> 19:                 Cogeneration:ElectricityNet     J
#> 20:                  Carbon Equivalent:Facility    kg
#> 21: CarbonEquivalentEmissions:Carbon Equivalent    kg
#>                                        variable units

# }