Read an EnergyPlus Simulation Error File

read_err() takes a file path of EnergyPlus simulation error file, usually with an extension .err, parses it and returns an ErrFile object.

read_err(path)

Arguments

path
a file path of EnergyPlus simulation error file, usually with an extension .err.

Value

An ErrFile object.

Details

Basically, an ErrFile object is a data.table with 6 columns and 6 additional attributes:

6 Columns:

index: Integer. Index of messages.

envir_index: Integer. Index of simulation environments.

envir: Character. Names of simulation environments.

level_index: Integer. Index for each severe level.

level: Character. Name of severe levels. Possible values: Info, Warning, Severe, and etc.

message: Character. Error messages.

6 Attributes:

path: A single string. The path of input file.

eplus_version: A numeric_version object. The version of EnergyPlus used during the simulation.

eplus_build: A single string. The build tag of EnergyPlus used during the simulation.

datetime: A DateTime (POSIXct). The time when the simulation started.

idd_version: A numeric_version. The version of IDD used during the simulation.

successful: TRUE when the simulation ended successfully, and FALSE otherwise.

terminated: TRUE when the simulation was terminated, and FALSE otherwise.

Examples

# \dontrun{
# run simulation and get the err file
idf_name <- "1ZoneUncontrolled.idf"
epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
idf_path <- file.path(eplus_config(8.8)$dir, "ExampleFiles", idf_name)
epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData", epw_name)
job <- eplus_job(idf_path, epw_path)
#> Adding an object in class 'Output:SQLite' and setting its 'Option Type' to 'SimpleAndTabular' in order to create SQLite output file.
job$run(dir = tempdir())
#> Replace the existing IDF located at /tmp/Rtmpyc9nDP/1ZoneUncontrolled.idf.
#> ExpandObjects Started.
#> No expanded file generated.
#> ExpandObjects Finished. Time:     0.003
#> EnergyPlus Starting
#> EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2021.11.17 09:47
#> Processing Data Dictionary
#> Processing Input File
#> Initializing Simulation
#> Reporting Surfaces
#> Beginning Primary Simulation
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Warming up {19}
#> Warming up {20}
#> Warming up {21}
#> Warming up {22}
#> Starting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Starting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Starting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=01/21
#> Continuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=02/10
#> Continuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/02
#> Continuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/22
#> Continuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=04/11
#> Continuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/01
#> Continuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/21
#> Continuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/10
#> Continuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/30
#> Continuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=07/20
#> Continuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/09
#> Continuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/29
#> Continuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=09/18
#> Continuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/08
#> Continuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/28
#> Continuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=11/17
#> Continuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/07
#> Continuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/27
#> Continuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Writing tabular output file results using comma format.
#> Writing tabular output file results using tab format.
#> Writing tabular output file results using text format.
#> Writing tabular output file results using HTML format.
#> Writing tabular output file results using XML format.
#> Writing final SQL reports
#> EnergyPlus Run Time=00hr 00min  2.29sec
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.13sec
#>  ReadVarsESO program completed successfully.
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.26sec
#>  ReadVarsESO program completed successfully.
#> ── EnergPlus Simulation Job ────────────────────────────────────────────────────
#> * Model: '/tmp/Rtmpyc9nDP/1ZoneUncontrolled.idf'
#> * Weather: '/home/runner/.local/EnergyPlus-8-8-0/WeatherData/USA_CA_San.Fra...
#> * EnergyPlus Version: '8.8.0'
#> * EnergyPlus Path: '/home/runner/.local/EnergyPlus-8-8-0'
#>  Simulation started at '2021-11-17 09:47:21' and completed successfully after 2.89 secs.

# read the err file
read_err(job$locate_output(".err"))
#> ══ EnergyPlus Error File ═══════════════════════════════════════════════════════
#>   * EnergyPlus version: 8.8.0 (7c3bbe4830)
#>   * Simulation started: 2021-11-17 09:47:00
#>   * Terminated: FALSE
#>   * Successful: TRUE
#>   * Warning[W]: 3
#> 
#> ── During Simulation Initiation ────────────────────────────────────────────────
#> [W 1/3] Weather file location will be used rather than entered (IDF) Location
#>         object.
#>         ..Location object=DENVER CENTENNIAL GOLDEN N_CO_USA DESIGN_CONDITIONS
#>         ..Weather File Location=San Francisco Intl Ap CA USA TMY3 WMO#=724940
#>         ..due to location differences, Latitude difference=[2.12] degrees,
#>         Longitude difference=[17.22] degrees.
#>         ..Time Zone difference=[1.0] hour(s), Elevation difference=[99.89]
#>         percent, [1827.00] meters.
#> [W 2/3] SetUpDesignDay: Entered DesignDay Barometric Pressure=81198 differs by
#>         more than 10% from Standard Barometric Pressure=101301.
#>         ...occurs in DesignDay=DENVER CENTENNIAL GOLDEN N ANN HTG 99% CONDNS DB,
#>         Standard Pressure (based on elevation) will be used.
#> [W 3/3] SetUpDesignDay: Entered DesignDay Barometric Pressure=81198 differs by
#>         more than 10% from Standard Barometric Pressure=101301.
#>         ...occurs in DesignDay=DENVER CENTENNIAL GOLDEN N ANN CLG 1% CONDNS
#>         DB=>MWB, Standard Pressure (based on elevation) will be used.
#> [I 1/8] Testing Individual Branch Integrity
#> [I 2/8] All Branches passed integrity testing
#> [I 3/8] Testing Individual Supply Air Path Integrity
#> [I 4/8] All Supply Air Paths passed integrity testing
#> [I 5/8] Testing Individual Return Air Path Integrity
#> [I 6/8] All Return Air Paths passed integrity testing
#> [I 7/8] No node connection errors were found.
#> 
#> ── During Simulation ───────────────────────────────────────────────────────────
#> [I 8/8] Simulation Error Summary *************
# }