Retreive Simulation Outputs Using EnergyPlus SQLite Output File

EplusSql class wraps SQL queries that can retreive simulation outputs using EnergyPlus SQLite output file.

Details

SQLite output is an optional output format for EnergyPlus. It will be created if there is an object in class Output:SQLite. If the value of field Option in class Output:SQLite is set to "SimpleAndTabular", then database tables related to the tabular reports will be included.

There are more than 30 tables in the SQLite output file which contains all of the data found in EnergyPlus's tabular output files, standard variable and meter output files, plus a number of reports that are found in the eplusout.eio output file. The full description for SQLite outputs is described in the EnergyPlus "Output Details and Examples" documentation.

EplusSql class makes it possible to directly retreive simulation results without creating an EplusJob object which can only get simulation outputs after the job was successfully run before.

However, it should be noted that, unlike EplusJob, there is no checking on whether the simulation is terminated or completed unsuccessfully or, the parent Idf has been changed since last simulation. This means that you may encounter some problems when retreive data from an unsuccessful simulation. It is suggested to carefully go through the .err file to make sure the output data in the SQLite is correct and reliable.

Usage

epsql <- eplus_sql(sql)
epsql$path()
epsql$list_table()
epsql$read_table(table)
epsql$report_data(key_value = NULL, name = NULL, year = NULL, tz = "GMT", case = "auto", all = FALSE)
epsql$report_data_dict()
epsql$tabular_data()
epsql$print()
print(epsql)

$path() returns the path of EnergyPlus SQLite file.

$report_data_dict() returns a data.table which contains all information about report data. For details on the meaning of each columns, please see "2.20.2.1 ReportDataDictionary Table" in EnergyPlus "Output Details and Examples" documentation.

$report_data() extracts the report data in a data.table using key values and variable names.

$tabular_data() extracts all tabular data in a data.table.

$print() shows the core information of this EplusSql object, including the path of the EnergyPlus SQLite file, last modified time of the SQLite file and the path of the IDF file with the same name in the same folder.

Arguments:

epsql: An EplusSQL object.

sql: A path to an local EnergyPlus SQLite output file.

key_value: A character vector to identify key name of the data. If NULL, all keys of that variable will be returned. Default: NULL.

name: A character vector to specify the actual data name. If NULL, all variables will be returned. Default: NULL.

year: The year of the date and time in column DateTime. If NULL, it will be the current year. Default: NULL

tz: Time zone of date and time in column DateTime. Default: "GMT".

case: If not NULL, a character column will be added indicates the case of this simulation. If "auto", the name of the SQL file will be used.

Examples

if (is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData", epw_name)

    # copy to tempdir and run the model
    idf <- read_idf(idf_path)
    idf$run(epw_path, tempdir())

    # create from local file
    sql <- eplus_sql(file.path(tempdir(), "1ZoneUncontrolled.sql"))

    # get sql file path
    sql$path()

    # list all tables in the sql file
    sql$list_table()

    # read a specific table
    sql$read_table("Zones")

    # read report data dictionary
    sql$report_data_dict()

    # read report data
    sql$report_data(name = "EnergyTransfer:Building")

    # read tabular data
    sql$tabular_data()
}

#> Idd v8.8.0 has not been parsed before. Try to locate `Energy+.idd` in EnergyPlus v8.8.0 installation folder `/usr/local/EnergyPlus-8-8-0`.

#> IDD file found: `/usr/local/EnergyPlus-8-8-0/Energy+.idd`.

#> Start parsing...

#> ── Info ────────────────────────────────────────────────────────────────────────
#> Adding object `Output:SQLite` and setting `Option Type` to `SimpleAndTabular`.
#> 
#> ── Info ────────────────────────────────────────────────────────────────────────
#> Replace the existing file located  at /tmp/Rtmp9PDt2A/1ZoneUncontrolled.idf.
#> 
#> ExpandObjects Started.
#> No expanded file generated.
#> ExpandObjects Finished. Time:     0.004
#> EnergyPlus Starting
#> EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2018.09.04 13:46
#> Processing Data Dictionary
#> Processing Input File
#> Initializing Simulation
#> Reporting Surfaces
#> Beginning Primary Simulation
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Warming up {19}
#> Warming up {20}
#> Warming up {21}
#> Warming up {22}
#> Starting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Starting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Starting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=01/21
#> Continuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=02/10
#> Continuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/02
#> Continuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/22
#> Continuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=04/11
#> Continuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/01
#> Continuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/21
#> Continuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/10
#> Continuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/30
#> Continuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=07/20
#> Continuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/09
#> Continuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/29
#> Continuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=09/18
#> Continuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/08
#> Continuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/28
#> Continuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=11/17
#> Continuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/07
#> Continuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/27
#> Continuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Writing tabular output file results using comma format.
#> Writing tabular output file results using tab format.
#> Writing tabular output file results using text format.
#> Writing tabular output file results using HTML format.
#> Writing tabular output file results using XML format.
#> Writing final SQL reports
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.27sec
#>  ReadVarsESO program completed successfully.
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.12sec
#>  ReadVarsESO program completed successfully.
#> EnergyPlus Run Time=00hr 00min  2.84sec
#> 

#>       TabularDataIndex        Value                              ReportName
#>    1:                1        82.43 AnnualBuildingUtilityPerformanceSummary
#>    2:                2        82.43 AnnualBuildingUtilityPerformanceSummary
#>    3:                3       261.06 AnnualBuildingUtilityPerformanceSummary
#>    4:                4       261.06 AnnualBuildingUtilityPerformanceSummary
#>    5:                5       354.92 AnnualBuildingUtilityPerformanceSummary
#>   ---                                                                      
#> 2154:             2154         Pass                  Initialization Summary
#> 2155:             2155         Pass                  Initialization Summary
#> 2156:             2156         Pass                  Initialization Summary
#> 2157:             2157         Pass                  Initialization Summary
#> 2158:             2158         Pass                  Initialization Summary
#>       ReportForString                      TableName             RowName
#>    1: Entire Facility         Site and Source Energy   Total Site Energy
#>    2: Entire Facility         Site and Source Energy     Net Site Energy
#>    3: Entire Facility         Site and Source Energy Total Source Energy
#>    4: Entire Facility         Site and Source Energy   Net Source Energy
#>    5: Entire Facility         Site and Source Energy   Total Site Energy
#>   ---                                                                   
#> 2154: Entire Facility Warmup Convergence Information                   2
#> 2155: Entire Facility Warmup Convergence Information                   3
#> 2156: Entire Facility Warmup Convergence Information                   1
#> 2157: Entire Facility Warmup Convergence Information                   2
#> 2158: Entire Facility Warmup Convergence Information                   3
#>                               ColumnName Units
#>    1:                       Total Energy    GJ
#>    2:                       Total Energy    GJ
#>    3:                       Total Energy    GJ
#>    4:                       Total Energy    GJ
#>    5:     Energy Per Total Building Area MJ/m2
#>   ---                                         
#> 2154: Heating Load Pass/Fail Convergence      
#> 2155: Heating Load Pass/Fail Convergence      
#> 2156: Cooling Load Pass/Fail Convergence      
#> 2157: Cooling Load Pass/Fail Convergence      
#> 2158: Cooling Load Pass/Fail Convergence