Read, and modify an EnergyPlus Weather File (EPW)

Reading an EPW file starts with function read_epw(), which parses an EPW file and returns an Epw object. The parsing process is extremely inspired by OpenStudio utilities library with some simplifications.

Details

An EPW file can be divided into two parts, headers and weather data. The first eight lines of a standard EPW file are normally headers which contains data of location, design conditions, typical/extreme periods, ground temperatures, holidays/daylight savings, data periods and other comments. For now, eplusr only parses headers of location, holidays/daylight savings and data periods. All other headers are left as they were when parsing and saving. For details on the data structure of EPW file, please see "Chapter 2 - Weather Converter Program" in EnergyPlus "Auxiliary Programs" documentation. An online version can be found here.

There are about 35 variables in the core weather data. However, not all of them are used by EnergyPlus. Actually, despite of date and time columns, only 14 columns are used:

dry bulb temperature

dew point temperature

relative humidity

atmospheric pressure

horizontal infrared radiation intensity from sky

direct normal radiation

diffuse horizontal radiation

wind direction

wind speed

present weather observation

present weather codes

snow depth

liquid precipitation depth

liquid precipitation rate

NOTE: Even though Epw class provides methods to replace core weather data, it is still not recommended.

Usage

epw <- read_epw(path)
epw$city
epw$city <- "city"
epw$state_province
epw$state_province <- "state_province"
epw$country
epw$country <- "country"
epw$data_source
epw$data_source <- "data_source"
epw$wmo_number
epw$wmo_number <- "wmo_number"
epw$latitude
epw$latitude <- "latitude"
epw$longitute
epw$longitute <- "longitute"
epw$time_zone
epw$time_zone <- "time_zone"
epw$elevation
epw$elevation <- "elevation"
epw$time_step
epw$time_step <- "time_step"
epw$start_day_of_week
epw$start_day_of_week <- "start_day_of_week"
epw$path()
epw$get_data(year = NULL, unit = FALSE, tz = Sys.timezone(), update = FALSE)
epw$set_data(data)
epw$save(path, overwrite = FALSE)
epw$print()
print(epw)

Read

epw <- read_epw(path)

Arguments

path: Path of an EnergyPlus EPW file.

epw$save(path, overwrite = FALSE)

Arguments

path: A path where to save the weather file. If NULL, the path of the weather file itself is used.

overwrite: Whether to overwrite the file if it already exists. Default is FALSE.

Query and Modify Header

epw$city
epw$city <- "city"
epw$state_province
epw$state_province <- "state_province"
epw$country
epw$country <- "country"
epw$data_source
epw$data_source <- "data_source"
epw$wmo_number
epw$wmo_number <- "wmo_number"
epw$latitude
epw$latitude <- "latitude"
epw$longitute
epw$longitute <- "longitute"
epw$time_zone
epw$time_zone <- "time_zone"
epw$elevation
epw$elevation <- "elevation"
epw$time_step
epw$time_step <- "time_step"
epw$start_day_of_week
epw$start_day_of_week <- "start_day_of_week"

$city, $state_province, $country, $data_source, $wmo_number, $latitude, $longitute, $time_zone, $elevation, $time_step and $start_day_of_week are all active bindings, which means that you can get the value and also set new value to it.

NOTE: Please be super careful when trying to modify those data. Some of them must be consistent with weather data in order to make the weather file successfully parsed by EnergyPlus and eplusr.

Query and Modify Data

epw$path()
epw$get_data(year = NULL, unit = FALSE, tz = Sys.timezone(), update = FALSE)
epw$set_data(data)

$path() returns the path of EPW file. NULL if Epw is not created from local file.

$get_data() returns the core weather data in a data.table.

$set_data() replaces core weather data with input data. NOTE: This feature is experimental. There is no validation when replacing.

Arguments

year: A integer to indicate the year value in the return datetime column. If NULL, which is the default, the year is left as it is in EPW file.

tz: The time zone of Date and Time in datetime column. Default: value of Sys.timezone().

unit: If TRUE, units are set to all numeric columns using units::set_units(). Default: FALSE.

update: If TRUE, not only datetime column, but also year, month, day, hour and minute are also updated according to the input year value. Default: FALSE

data: A data.frame which has all required columns.

Examples

# read an EPW file from EnergyPlus website
path_base <- "https://energyplus.net/weather-download"
path_region <- "north_and_central_america_wmo_region_4/USA/CA"
path_file <- "USA_CA_San.Francisco.Intl.AP.724940_TMY3/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
path_epw <- file.path(path_base, path_region, path_file)
epw <- read_epw(path_epw)

# read an EPW file distributed with EnergyPlus
if (is_avail_eplus(8.8)) {
    epw_path <- file.path(
        eplus_config(8.8)$dir,
        "WeatherData",
        "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
    )
    epw <- read_epw(path_epw)
}

# get file path
epw$path()

#> NULL

# get basic info
epw$city

#> [1] "San Francisco Intl Ap"

epw$state_province

#> [1] "CA"

epw$country

#> [1] "USA"

epw$data_source

#> [1] "TMY3"

epw$wmo_number

#> [1] "724940"

epw$latitude

#> [1] 37.62

epw$longitude

#> [1] -122.4

epw$time_zone

#> [1] -8

epw$elevation

#> [1] 2

epw$time_step

#> [1] 1

epw$start_day_of_week

#> [1] "Sunday"

# set basic info
# NOTE: Use with caution. May mess up your weather data
epw$city <- "Chongqing"
epw$city

#> [1] "Chongqing"

epw$state_province <- "Chongqing"
epw$state_province

#> [1] "Chongqing"

epw$country <- "China"
epw$country

#> [1] "China"

epw$data_source <- "TMY"
epw$data_source

#> [1] "TMY"

epw$wmo_number <- "724944"
epw$wmo_number

#> [1] "724944"

epw$latitude <- 20.0
epw$latitude

#> [1] 20

epw$longitude <- -120.0
epw$longitude

#> [1] -120

epw$time_zone <- 8
epw$time_zone

#> [1] 8

epw$elevation <- 100
epw$elevation

#> [1] 100

epw$time_step <- 2
epw$time_step

#> [1] 2

epw$start_day_of_week <- "Monday"
epw$start_day_of_week

#> [1] "Monday"

# get weather data
str(epw$get_data())

#> Classes ‘data.table’ and 'data.frame':	8760 obs. of  36 variables:
#>  $ datetime                                        : POSIXct, format: "1999-01-01 01:00:00" "1999-01-01 02:00:00" ...
#>  $ year                                            : int  1999 1999 1999 1999 1999 1999 1999 1999 1999 1999 ...
#>  $ month                                           : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ day                                             : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ hour                                            : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ minute                                          : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ datasource                                      : chr  "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" ...
#>  $ dry_bulb_temperature                            : num  7.2 7.2 6.7 6.1 4.4 4.4 6.1 5 7.8 8.9 ...
#>  $ dew_point_temperature                           : num  5.6 5.6 5 5 3.9 3.9 4.4 2.8 5 6.7 ...
#>  $ relative_humidity                               : num  90 90 89 93 97 97 89 86 82 86 ...
#>  $ atmospheric_pressure                            : num  102200 102100 102200 102200 102200 ...
#>  $ extraterrestrial_horizontal_radiation           : num  0 0 0 0 0 0 0 36 243 443 ...
#>  $ extraterrestrial_direct_normal_radiation        : num  0 0 0 0 0 ...
#>  $ horizontal_infrared_radiation_intensity_from_sky: num  290 296 291 276 280 280 287 281 295 299 ...
#>  $ global_horizontal_radiation                     : num  0 0 0 0 0 0 0 2 121 276 ...
#>  $ direct_normal_radiation                         : num  0 0 0 0 0 0 0 69 477 651 ...
#>  $ diffuse_horizontal_radiation                    : num  0 0 0 0 0 0 0 1 39 71 ...
#>  $ global_horizontal_illuminance                   : num  0 0 0 0 0 ...
#>  $ direct_normal_illuminance                       : num  0 0 0 0 0 ...
#>  $ diffuse_horizontal_illuminance                  : num  0 0 0 0 0 ...
#>  $ zenith_luminance                                : num  0 0 0 0 0 ...
#>  $ wind_direction                                  : num  0 170 210 200 260 180 0 0 0 0 ...
#>  $ wind_speed                                      : num  0 2.1 2.1 1.5 3.1 2.1 0 0 0 0 ...
#>  $ total_sky_cover                                 : int  2 4 3 0 3 3 3 3 3 2 ...
#>  $ opaque_sky_cover                                : int  2 4 3 0 3 3 3 3 3 2 ...
#>  $ visibility                                      : num  16 16 16 16 16 16 16 16 16 16 ...
#>  $ ceiling_height                                  : num  77777 77777 77777 77777 77777 ...
#>  $ present_weather_observation                     : int  9 9 9 9 9 9 9 9 9 9 ...
#>  $ present_weather_codes                           : chr  NA NA NA NA ...
#>  $ precipitable_water                              : num  129 120 120 120 120 120 120 110 110 110 ...
#>  $ aerosol_optical_depth                           : num  0.108 0.108 0.108 0.108 0.108 0.108 0.108 0.108 0.108 0.108 ...
#>  $ snow_depth                                      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ days_since_last_snow                            : int  88 88 88 88 88 88 88 88 88 88 ...
#>  $ albedo                                          : num  0.16 0.16 0.16 0.16 0.16 0.16 0.16 0.16 0.16 0.16 ...
#>  $ liquid_precip_depth                             : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liquid_precip_rate                              : num  1 1 1 1 1 1 1 1 1 1 ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

# get weather data but change the year to 2018
# the year column is not changed by default, only the returned datetime column
str(epw$get_data(year = 2018)$datetime)

#>  POSIXct[1:8760], format: "2018-01-01 01:00:00" "2018-01-01 02:00:00" "2018-01-01 03:00:00" ...

str(epw$get_data(year = 2018)$year)

#>  int [1:8760] 1999 1999 1999 1999 1999 1999 1999 1999 1999 1999 ...

# you can force to update the year column
str(epw$get_data(year = 2018, update = TRUE)$year)

#>  int [1:8760] 2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 ...

# get weather data with units
str(epw$get_data(unit = TRUE))

#> Classes ‘data.table’ and 'data.frame':	8760 obs. of  36 variables:
#>  $ datetime                                        : POSIXct, format: "1999-01-01 01:00:00" "1999-01-01 02:00:00" ...
#>  $ year                                            : int  1999 1999 1999 1999 1999 1999 1999 1999 1999 1999 ...
#>  $ month                                           : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ day                                             : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ hour                                            : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ minute                                          : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ datasource                                      : chr  "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" "?9?9?9?9E0?9?9?9?9?9?9?9?9?9?9?9?9?9?9?9*9*9?9?9?9" ...
#>  $ dry_bulb_temperature                            :Object of class units:
#>  num  7.2 7.2 6.7 6.1 4.4 4.4 6.1 5 7.8 8.9 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "°C"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ dew_point_temperature                           :Object of class units:
#>  num  5.6 5.6 5 5 3.9 3.9 4.4 2.8 5 6.7 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "°C"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ relative_humidity                               :Object of class units:
#>  num  90 90 89 93 97 97 89 86 82 86 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "%"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ atmospheric_pressure                            :Object of class units:
#>  num  102200 102100 102200 102200 102200 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "Pa"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ extraterrestrial_horizontal_radiation           :Object of class units:
#>  num  0 0 0 0 0 0 0 36 243 443 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr  "h" "W"
#>   .. ..$ denominator: chr  "m" "m"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ extraterrestrial_direct_normal_radiation        :Object of class units:
#>  num  0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr  "h" "W"
#>   .. ..$ denominator: chr  "m" "m"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ horizontal_infrared_radiation_intensity_from_sky:Object of class units:
#>  num  290 296 291 276 280 280 287 281 295 299 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr  "h" "W"
#>   .. ..$ denominator: chr  "m" "m"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ global_horizontal_radiation                     :Object of class units:
#>  num  0 0 0 0 0 0 0 2 121 276 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr  "h" "W"
#>   .. ..$ denominator: chr  "m" "m"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ direct_normal_radiation                         :Object of class units:
#>  num  0 0 0 0 0 0 0 69 477 651 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr  "h" "W"
#>   .. ..$ denominator: chr  "m" "m"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ diffuse_horizontal_radiation                    :Object of class units:
#>  num  0 0 0 0 0 0 0 1 39 71 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr  "h" "W"
#>   .. ..$ denominator: chr  "m" "m"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ global_horizontal_illuminance                   :Object of class units:
#>  num  0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "lx"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ direct_normal_illuminance                       :Object of class units:
#>  num  0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "lx"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ diffuse_horizontal_illuminance                  :Object of class units:
#>  num  0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "lx"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ zenith_luminance                                :Object of class units:
#>  num  0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "lx"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ wind_direction                                  :Object of class units:
#>  num  0 170 210 200 260 180 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "°"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ wind_speed                                      :Object of class units:
#>  num  0 2.1 2.1 1.5 3.1 2.1 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "m"
#>   .. ..$ denominator: chr "s"
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ total_sky_cover                                 :Object of class units:
#>  int  2 4 3 0 3 3 3 3 3 2 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr 
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ opaque_sky_cover                                :Object of class units:
#>  int  2 4 3 0 3 3 3 3 3 2 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr 
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ visibility                                      :Object of class units:
#>  num  16 16 16 16 16 16 16 16 16 16 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "km"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ ceiling_height                                  :Object of class units:
#>  num  77777 77777 77777 77777 77777 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "m"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ present_weather_observation                     :Object of class units:
#>  int  9 9 9 9 9 9 9 9 9 9 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr 
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ present_weather_codes                           : chr  NA NA NA NA ...
#>  $ precipitable_water                              :Object of class units:
#>  num  129 120 120 120 120 120 120 110 110 110 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "mm"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ aerosol_optical_depth                           :Object of class units:
#>  num  0.108 0.108 0.108 0.108 0.108 0.108 0.108 0.108 0.108 0.108 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr 
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ snow_depth                                      :Object of class units:
#>  num  0 0 0 0 0 0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "cm"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ days_since_last_snow                            :Object of class units:
#>  int  88 88 88 88 88 88 88 88 88 88 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr 
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ albedo                                          :Object of class units:
#>  num  0.16 0.16 0.16 0.16 0.16 0.16 0.16 0.16 0.16 0.16 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr 
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ liquid_precip_depth                             :Object of class units:
#>  num  0 0 0 0 0 0 0 0 0 0 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "mm"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  $ liquid_precip_rate                              :Object of class units:
#>  num  1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, "units")=List of 2
#>   .. ..$ numerator  : chr "h"
#>   .. ..$ denominator: chr 
#>   .. ..- attr(*, "class")= chr "symbolic_units"
#>  - attr(*, ".internal.selfref")=<externalptr> 

# with units specified, you can easily perform unit conversion using units
# package
t_dry_bulb <- epw$get_data(unit = TRUE)$dry_bulb_temperature
units(t_dry_bulb) <- with(units::ud_units, "kelvin")
str(t_dry_bulb)

#> Object of class units:
#>  num [1:8760] 280 280 280 279 278 ...
#>  - attr(*, "units")=List of 2
#>   ..$ numerator  : chr "K"
#>   ..$ denominator: chr(0) 
#>   ..- attr(*, "class")= chr "symbolic_units"

# change the time zone of datetime column in the returned weather data
attributes(epw$get_data()$datetime)

#> $class
#> [1] "POSIXct" "POSIXt" 
#> 
#> $tzone
#> [1] "Asia/Shanghai"
#> 

attributes(epw$get_data(tz = "America/Chicago")$datetime)

#> $class
#> [1] "POSIXct" "POSIXt" 
#> 
#> $tzone
#> [1] "America/Chicago"
#> 

# NOT RUN {
# change the weather data
# NOTE: This feature is experimental. There is no validation when replacing.
epw$set_data(epw$get_data())
# }

# save the weather file
epw$save(file.path(tempdir(), "weather.epw"))