EnergyPlus IDF object

IdfObject is an abstraction of a single object in an Idf. It provides more detail methods to modify objects. IdfObject can only be created from the parent Idf object, using $object, $object_in_class() and $search_object() or equivalent. This is because that initialization of an IdfObject needs some shared data from parent Idf object.

Usage

idfobj$id()
idfobj$name()
idfobj$definition()
idfobj$get_comment()
idfobj$set_comment(comment, append = TRUE, width = 0L)
idfobj$get_value(which = NULL, all = NULL, simplify = FALSE)
idfobj$set_value(..., defaults = TRUE)
idfobj$FieldName
idfobj[[Field]]
idfobj$FieldName <- Value
idfobj[[Field]] <- Value
idfobj$possible_value(which = NULL)
idfobj$validate()
idfobj$is_valid()
idfobj$ref_from_object()
idfobj$ref_by_object()
idfobj$has_ref_by()
idfobj$has_ref_from()
idfobj$has_ref()
idfobj$table(all = FALSE, unit = TRUE, wide = FALSE, string_value = TRUE, in_ip = eplusr_option("view_in_ip"))
idfobj$string(comment = TRUE, leading = 4L, sep_at = 29L)
idfobj$print(comment = TRUE, auto_sep = FALSE)
print(iddobj)

Basic Info

idfobj$id()
idfobj$name()
idfobj$group_name()
idfobj$class_name()

$id() returns the object ID.

$name() returns the object name. If the class does not have name attribute, then NA will returned.

$group_name() returns the group name of this object belongs to.

$class_name() returns the class name of this object belongs to.

Definition

idfobj$definition()

$definition() returns the definition, i.e. the IddObject, of current class. For details of IddObject, please see IddObject class.

Comment

idfobj$get_comment()
idfobj$set_comment(comment, append = TRUE, width = 0L)

$get_comment() returns the comments of current object.

$set_comment() sets comments of current object.

Arguments

comment: A character vector. If NULL, all comments will be deleted.

append: If TRUE, comment will be appended to existing comments. If FALSE, comment will be prepended to existing currents. If NULL, existing comments will be deleted before adding new comments. Default: FALSE.

width: An integer of character number to indicate where to break long comment lines. If 0, no breaking will be made. Default: 0.

Value

idfobj$get_value(which = NULL, all = FALSE, simplify = FALSE)
idfobj$set_value(..., default = TRUE)
idfobj$FieldName
idfobj[[Field]]
idfobj$FieldName <- Value
idfobj[[Field]] <- Value
idfobj$possible_value(which = NULL)

$get_value() returns a named list containing values of specified fields. If simplify is FALSE, then all values will be converted into character and the converted character vector will be returned. Note that the field names will be converted into valid R names, i.e. all characters other than letters and numbers will be replaced by underscore "_" ("underscore-style").

$set_value() sets values of current object. Field values should be given in following either pattern below:

directly list all field values with no name. The values will be assigned to fields according to the appearance order

give both field names without units and values in pair, e.g. Name = "Test", `Begin Month` = 1. You can find all valid field names using $definition()$field_name(). Field names can also be given in underscore-style, e.g. Name = "Test", begin_month = 1 (NOTE: matching is case-insensitive).

eplusr also provides custom S3 method of $, [[ and also $<- and [[<- to make it more convenient to get and set a single value of an IdfObject. Basically, idfobj$FieldName and idfobj[[Field is equivalent to idfobj$get_value(Field)[[1]]; idfobj$FieldName <- Value and idfobj[[Field]] <- Value is equivalent to idfobj$set_value(Field = Value), where FieldName is a single valid field name and Field is a single valid field index or name.

$possible_value() return all possible values for specified fields, including auto-value (autosize and autocalculate), default value, value range, choices and references. Underneath, it returns a data.table with custom printing method. It is basically the same as $field_possible() in IddObject class.

Arguments

which: An integer vector of field indexes or a character vector of field names. Field names can be given in "lower-style", e.g. "Thermal Resistance" can be given as "thermal_resistance".

all: If TRUE, values of all fields, including empty fields will be returned as well. Default: FALSE

simplify: If TRUE, values of fields will be converted into characters and the converted character vector will be returned.

...: Values to set. Field names of value can be given. If not named, the input values will be set to fields according to their order of appearance.

default: If TRUE, all empty fields will be filled with their default values if possible.

FieldName: A single length character vector of one valid field name where all characters except letters and numbers are replaced by underscores.

Field: A single length character vector of one valid field name or a single length integer vector of one valid field index. Same as above, field names should be given in a style where all characters except letters and numbers are replaced by underscores.

Value: A single length vector of value to set.

Validation

idfobj$validate()
idfobj$is_valid()

$validate() will check if there are errors in current object under different strictness level.

$is_valid() will check if there are no errors in current object under different strictness level.

The strictness level can be changed using eplusr_option(). Default is "final".There are three different validate levels, i.e. "none", "draft" and "final":

For "none", none validation will be done;

For "draft", checking of invalid autosize, autocalculate, character, numeric, integer, and choice field values will be done;

For "final", besides above, checking of incomplete extensible groups, missing required objects, duplicated unique objects, object name conflicts, missing required fields and invalid field value reference will also be done.

For details about the underlying structure of returned value of $validate(), please $validate() in Idf class.

Cross Reference

idfobj$ref_from_object()
idfobj$ref_by_object()
idfobj$has_ref_from()
idfobj$has_ref_by()
idfobj$has_ref()

$ref_from_object() will return other objects that current object references from.

$ref_by_object() will return other objects that reference current object.

$has_ref_from() and $has_ref_by will return TRUE if current object has referenced from other objects or has been referenced by other objects, respectively.

$has_ref() will return TRUE if current object has either referenced from other objects or has been referenced by other objects.

Data Extraction

idfobj$table(all = FALSE, unit = TRUE, wide = FALSE, string_value = TRUE, in_ip = eplusr_option("view_in_ip"))

$table() will return a data.table that contains all data of current object.

Arguments

all: If TRUE, values of all fields, including empty fields will be returned as well. Default: FALSE

unit: If TRUE, field names with units will be returned. Default: TRUE.

wide: If TRUE, a wide table will be returned. Default: FALSE.

string_value: If TRUE, all field values will be returned as character. Default: TRUE

in_ip: If TRUE, IP units and values will be returned. Default: the value of eplusr_option("view_in_ip").

Formatting

idfobj$string(comment = TRUE, leading = 4L, sep_at = 29L)

$string() will return the text format of current object.

Arguments

comment: If FALSE, all comments will not be included.

leading: An integer to indicate the number of spaces before each fields. Default: 4.

sep_at: An integer to indicate the character width where to separate values and field names. Default: 29.

Print

idfobj$print(comment = TRUE, auto_sep = FALSE)
print(idfobj)

Arguments

comment: If FALSE, all comments will not be included.

auto_sep: If TRUE, values and field names will be separate at the largest character length of values. Default: FALSE.

See also

Examples

# read an IDF file
idf <- read_idf(system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr"),
    idd = use_idd(8.8, download = "auto"))

# get the IdfObject of material named "C5 - 4 IN HW CONCRETE"
mat <- idf$Material$C5_4_IN_HW_CONCRETE

# get object ID
mat$id()

#> [1] 14

# get object name
mat$name()

#> [1] "C5 - 4 IN HW CONCRETE"

# NA will be returned if the class does not have name attribute. For example,
# "Version" class
idf$Version[[1]]$name()

#> [1] NA

# get underlying IddObject of current class
mat$definition()

#> << Class: `Material` >>
#> ─────────────────────────────────── * MEMO * ───────────────────────────────────
#>   "Regular materials described with full set of thermal properties"
#> ──────────────────────────────── * PROPERTIES * ────────────────────────────────
#>   Group: `Surface Construction Elements`
#>   Unique: FALSE
#>   Required: FALSE
#>   Total fields: 9
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#>   1:* Name
#>   2:* Roughness
#>   3:* Thickness {m}
#>   4:* Conductivity {W/m-K}
#>   5:* Density {kg/m3}
#>   6:* Specific Heat {J/kg-K}
#>   7:  Thermal Absorptance
#>   8:  Solar Absorptance
#>   9:  Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

# get object comments
mat$get_comment()

#> character(0)

# add new object comments
mat$set_comment(c("This is a material named `WD01`", "This object has an ID of 47"))

#> <<[ID:14] `C5 - 4 IN HW CONCRETE`>> Material
#> ───────────────────────────────── * COMMENTS * ─────────────────────────────────
#> !This is a material named `WD01`
#> !This object has an ID of 47
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: C5 - 4 IN HW CONCRETE,  !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.1014984,         !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat$get_comment()

#> [1] "This is a material named `WD01`" "This object has an ID of 47"    

# append new comments
mat$set_comment("This is an appended comment")

#> <<[ID:14] `C5 - 4 IN HW CONCRETE`>> Material
#> ───────────────────────────────── * COMMENTS * ─────────────────────────────────
#> !This is a material named `WD01`
#> !This object has an ID of 47
#> !This is an appended comment
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: C5 - 4 IN HW CONCRETE,  !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.1014984,         !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat$get_comment()

#> [1] "This is a material named `WD01`" "This object has an ID of 47"    
#> [3] "This is an appended comment"    

# prepend new comments
mat$set_comment("This is a prepended comment", append = FALSE)

#> <<[ID:14] `C5 - 4 IN HW CONCRETE`>> Material
#> ───────────────────────────────── * COMMENTS * ─────────────────────────────────
#> !This is a prepended comment
#> !This is a material named `WD01`
#> !This object has an ID of 47
#> !This is an appended comment
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: C5 - 4 IN HW CONCRETE,  !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.1014984,         !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat$get_comment()

#> [1] "This is a prepended comment"     "This is a material named `WD01`"
#> [3] "This object has an ID of 47"     "This is an appended comment"    

# wrap long comments
mat$set_comment("This is a very long comment that is needed to be wrapped.", width = 30)

#> <<[ID:14] `C5 - 4 IN HW CONCRETE`>> Material
#> ───────────────────────────────── * COMMENTS * ─────────────────────────────────
#> !This is a prepended comment
#> !This is a material named `WD01`
#> !This object has an ID of 47
#> !This is an appended comment
#> !This is a very long comment
#> !that is needed to be wrapped.
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: C5 - 4 IN HW CONCRETE,  !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.1014984,         !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat$get_comment()

#> [1] "This is a prepended comment"     "This is a material named `WD01`"
#> [3] "This object has an ID of 47"     "This is an appended comment"    
#> [5] "This is a very long comment"     "that is needed to be wrapped."  

# delete old comments and add new one
mat$set_comment("This is the only comment", append = NULL)

#> <<[ID:14] `C5 - 4 IN HW CONCRETE`>> Material
#> ───────────────────────────────── * COMMENTS * ─────────────────────────────────
#> !This is the only comment
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: C5 - 4 IN HW CONCRETE,  !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.1014984,         !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat$get_comment()

#> [1] "This is the only comment"

# delete all comments
mat$set_comment(NULL)

#> <<[ID:14] `C5 - 4 IN HW CONCRETE`>> Material
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: C5 - 4 IN HW CONCRETE,  !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.1014984,         !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat$get_comment()

#> character(0)

# get all existing field values
str(mat$get_value())

#> List of 9
#>  $ Name               : chr "C5 - 4 IN HW CONCRETE"
#>  $ Roughness          : chr "MediumRough"
#>  $ Thickness          : num 0.101
#>  $ Conductivity       : num 1.73
#>  $ Density            : num 2243
#>  $ Specific_Heat      : num 837
#>  $ Thermal_Absorptance: num 0.9
#>  $ Solar_Absorptance  : num 0.65
#>  $ Visible_Absorptance: num 0.65

# get values of field 1, 3, 5
str(mat$get_value(c(1, 3, 5)))

#> List of 3
#>  $ Name     : chr "C5 - 4 IN HW CONCRETE"
#>  $ Thickness: num 0.101
#>  $ Density  : num 2243

# get character format values instead of a named list
mat$get_value(c(1, 3, 5), simplify = TRUE)

#> [1] "C5 - 4 IN HW CONCRETE" "0.1014984"             "2242.585"             

# get values of all field even those that are not set
str(idf$Zone$ZONE_ONE$get_value())

#> List of 9
#>  $ Name                       : chr "ZONE ONE"
#>  $ Direction_of_Relative_North: num 0
#>  $ X_Origin                   : num 0
#>  $ Y_Origin                   : num 0
#>  $ Z_Origin                   : num 0
#>  $ Type                       : num 1
#>  $ Multiplier                 : num 1
#>  $ Ceiling_Height             : chr "autocalculate"
#>  $ Volume                     : chr "autocalculate"

str(idf$Zone$ZONE_ONE$get_value(all = TRUE))

#> List of 13
#>  $ Name                             : chr "ZONE ONE"
#>  $ Direction_of_Relative_North      : num 0
#>  $ X_Origin                         : num 0
#>  $ Y_Origin                         : num 0
#>  $ Z_Origin                         : num 0
#>  $ Type                             : num 1
#>  $ Multiplier                       : num 1
#>  $ Ceiling_Height                   : chr "autocalculate"
#>  $ Volume                           : chr "autocalculate"
#>  $ Floor_Area                       : num NA
#>  $ Zone_Inside_Convection_Algorithm : chr NA
#>  $ Zone_Outside_Convection_Algorithm: chr NA
#>  $ Part_of_Total_Floor_Area         : chr NA

# get field values using shortcuts
mat$Roughness

#> [1] "MediumRough"

mat[["Specific_Heat"]]

#> [1] 836.8

mat[c(1,2)]

#> $Name
#> [1] "C5 - 4 IN HW CONCRETE"
#> 
#> $Roughness
#> [1] "MediumRough"
#> 

mat[c("Name", "Density")]

#> $Name
#> [1] "C5 - 4 IN HW CONCRETE"
#> 
#> $Density
#> [1] 2242.585
#> 

# set field values
mat$set_value(name = "new_name", Thickness = 0.02)

#> <<[ID:14] `new_name`>> Material
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: new_name,          !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.02,              !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.65,              !- Solar Absorptance
#>  9: 0.65;              !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat[c("Name", "Thickness")]

#> $Name
#> [1] "new_name"
#> 
#> $Thickness
#> [1] 0.02
#> 

# When `default` argument is set to TRUE and input field values are empty, i.e.
# NA and NULL, the field values will be reset to defaults.
mat[c("Thermal Absorptance", "Solar Absorptance")]

#> $Thermal_Absorptance
#> [1] 0.9
#> 
#> $Solar_Absorptance
#> [1] 0.65
#> 

mat$set_value(visible_absorptance = NA, Solar_Absorptance = NA, default = TRUE)

#> <<[ID:14] `new_name`>> Material
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: new_name,          !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.02,              !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: 2242.585,          !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.7,               !- Solar Absorptance
#>  9: 0.7;               !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

mat[c("Visible Absorptance", "Solar Absorptance")]

#> $Visible_Absorptance
#> [1] 0.7
#> 
#> $Solar_Absorptance
#> [1] 0.7
#> 

# set field values using shortcuts
mat$Name <- "another_name"
mat$Name

#> [1] "another_name"

mat[["Thickness"]] <- 0.019
mat$Thickness

#> [1] 0.019

# check validate
mat$validate()

#>  ✔ No error found.

mat$is_valid()

#> [1] TRUE

# if we set density to a negative number
mat$definition()$field_range("Density")

#> [[1]]
#> (0, Inf)
#> 

eplusr_option(validate_level = "none") # have to set validate to "none" to do so

#> $validate_level
#> [1] "none"
#> 

mat$Density <- -1
eplusr_option(validate_level = "final") # change back to "final" validate level

#> $validate_level
#> [1] "final"
#> 

mat$is_valid()

#> [1] FALSE

# get other objects that this object refereces
mat$ref_from_object() # not referencing other objects

#> Object does not reference from any other object.

mat$has_ref_from()

#> [1] FALSE

# get other objects that reference this object
mat$ref_by_object() # referenced by construction "FLOOR"

#> 1 object found that reference the target object [ID: 14].

#> $FLOOR
#> <<[ID:16] `FLOOR`>> Construction
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: FLOOR,             !- Name
#> *2: another_name;      !- Outside Layer
#> ────────────────────────────────────────────────────────────────────────────────
#> 

names(mat$ref_by_object())

#> 1 object found that reference the target object [ID: 14].

#> [1] "FLOOR"

mat$has_ref_by()

#> [1] TRUE

# check if having any referenced objects or is referenced by other objects
mat$has_ref()

#> [1] TRUE

# get all object data in a data.table format without field units
str(mat$table(unit = FALSE))

#> Classes ‘data.table’ and 'data.frame':	9 obs. of  3 variables:
#>  $ index: num  1 2 3 4 5 6 7 8 9
#>  $ name : chr  "Name" "Roughness" "Thickness" "Conductivity" ...
#>  $ value: chr  "another_name" "MediumRough" "0.019" "1.729577" ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

# get all object data in a data.table format where all field values are put in a
# list column and field names without unit
str(mat$table(string_value = FALSE, unit = FALSE))

#> Classes ‘data.table’ and 'data.frame':	9 obs. of  3 variables:
#>  $ index: num  1 2 3 4 5 6 7 8 9
#>  $ name : chr  "Name" "Roughness" "Thickness" "Conductivity" ...
#>  $ value:List of 9
#>   ..$ : chr "another_name"
#>   ..$ : chr "MediumRough"
#>   ..$ : num 0.019
#>   ..$ : num 1.73
#>   ..$ : num -1
#>   ..$ : num 837
#>   ..$ : num 0.9
#>   ..$ : num 0.7
#>   ..$ : num 0.7
#>  - attr(*, ".internal.selfref")=<externalptr> 

# get all object data in a data.table format where all field values are put in a
# list column and all values are converted into IP units
str(mat$table(string_value = FALSE, in_ip = TRUE))

#> Classes ‘data.table’ and 'data.frame':	9 obs. of  3 variables:
#>  $ index: num  1 2 3 4 5 6 7 8 9
#>  $ name : chr  "Name" "Roughness" "Thickness {ft}" "Conductivity {Btu-in/h-ft2-F}" ...
#>  $ value:List of 9
#>   ..$ : chr "another_name"
#>   ..$ : chr "MediumRough"
#>   ..$ : num 0.0623
#>   ..$ : num 12
#>   ..$ : num -0.0624
#>   ..$ : num 0.2
#>   ..$ : num 0.9
#>   ..$ : num 0.7
#>   ..$ : num 0.7
#>  - attr(*, ".internal.selfref")=<externalptr> 

# get all object data in a data.table format, including tailing empty fields
str(idf$Zone$ZONE_ONE$table(all = TRUE))

#> Classes ‘data.table’ and 'data.frame':	13 obs. of  3 variables:
#>  $ index: num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ name : chr  "Name" "Direction of Relative North {deg}" "X Origin {m}" "Y Origin {m}" ...
#>  $ value: chr  "ZONE ONE" "0" "0" "0" ...
#>  - attr(*, ".internal.selfref")=<externalptr> 

# get all object data in a data.table format where each field becomes a column
str(mat$table(wide = TRUE))

#> Classes ‘data.table’ and 'data.frame':	1 obs. of  9 variables:
#>  $ Name                  : chr "another_name"
#>  $ Roughness             : chr "MediumRough"
#>  $ Thickness {m}         : chr "0.019"
#>  $ Conductivity {W/m-K}  : chr "1.729577"
#>  $ Density {kg/m3}       : chr "-1"
#>  $ Specific Heat {J/kg-K}: chr "836.8"
#>  $ Thermal Absorptance   : chr "0.9"
#>  $ Solar Absorptance     : chr "0.7"
#>  $ Visible Absorptance   : chr "0.7"
#>  - attr(*, ".internal.selfref")=<externalptr> 

# get string format object
mat$string()

#>  [1] ""                                                      
#>  [2] "Material,"                                             
#>  [3] "    another_name,            !- Name"                  
#>  [4] "    MediumRough,             !- Roughness"             
#>  [5] "    0.019,                   !- Thickness {m}"         
#>  [6] "    1.729577,                !- Conductivity {W/m-K}"  
#>  [7] "    -1,                      !- Density {kg/m3}"       
#>  [8] "    836.8,                   !- Specific Heat {J/kg-K}"
#>  [9] "    0.9,                     !- Thermal Absorptance"   
#> [10] "    0.7,                     !- Solar Absorptance"     
#> [11] "    0.7;                     !- Visible Absorptance"   
#> [12] ""                                                      

# get string format of object, and decrease the space between field values and
# field names
mat$string(sep_at = 15)

#>  [1] ""                                        
#>  [2] "Material,"                               
#>  [3] "    another_name,  !- Name"              
#>  [4] "    MediumRough,  !- Roughness"          
#>  [5] "    0.019,     !- Thickness {m}"         
#>  [6] "    1.729577,  !- Conductivity {W/m-K}"  
#>  [7] "    -1,        !- Density {kg/m3}"       
#>  [8] "    836.8,     !- Specific Heat {J/kg-K}"
#>  [9] "    0.9,       !- Thermal Absorptance"   
#> [10] "    0.7,       !- Solar Absorptance"     
#> [11] "    0.7;       !- Visible Absorptance"   
#> [12] ""                                        

# get string format of object, and decrease the leading space of field values
mat$string(leading = 0)

#>  [1] ""                                                      
#>  [2] "Material,"                                             
#>  [3] "another_name,                !- Name"                  
#>  [4] "MediumRough,                 !- Roughness"             
#>  [5] "0.019,                       !- Thickness {m}"         
#>  [6] "1.729577,                    !- Conductivity {W/m-K}"  
#>  [7] "-1,                          !- Density {kg/m3}"       
#>  [8] "836.8,                       !- Specific Heat {J/kg-K}"
#>  [9] "0.9,                         !- Thermal Absorptance"   
#> [10] "0.7,                         !- Solar Absorptance"     
#> [11] "0.7;                         !- Visible Absorptance"   
#> [12] ""                                                      

# print the object without comment
mat$print(comment = FALSE)

#> <<[ID:14] `another_name`>> Material
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: another_name,      !- Name
#> *2: MediumRough,       !- Roughness
#> *3: 0.019,             !- Thickness {m}
#> *4: 1.729577,          !- Conductivity {W/m-K}
#> *5: -1,                !- Density {kg/m3}
#> *6: 836.8,             !- Specific Heat {J/kg-K}
#>  7: 0.9,               !- Thermal Absorptance
#>  8: 0.7,               !- Solar Absorptance
#>  9: 0.7;               !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────

# print the object, and auto separate field values and field names at the
# largetst character length of field values
mat$print(auto_sep = TRUE)

#> <<[ID:14] `another_name`>> Material
#> ────────────────────────────────── * FIELDS * ──────────────────────────────────
#> *1: another_name,  !- Name
#> *2: MediumRough,   !- Roughness
#> *3: 0.019,         !- Thickness {m}
#> *4: 1.729577,      !- Conductivity {W/m-K}
#> *5: -1,            !- Density {kg/m3}
#> *6: 836.8,         !- Specific Heat {J/kg-K}
#>  7: 0.9,           !- Thermal Absorptance
#>  8: 0.7,           !- Solar Absorptance
#>  9: 0.7;           !- Visible Absorptance
#> ────────────────────────────────────────────────────────────────────────────────