Read an EnergyPlus Report Data Dictionary File

read_rdd() takes a file path of EnergyPlus Report Data Dictionary (RDD) file, parses it and returns a RddFile object. read_mdd() takes a file path of EnergyPlus Meter Data Dictionary (MDD) file, parses it and returns a MddFile object.

read_rdd(path)

read_mdd(path)

Arguments

path
For read_rdd(), a file path of EnergyPlus EnergyPlus Report Data Dictionary file with an extension .rdd. For read_mdd(), a file path of EnergyPlus EnergyPlus Meter Data Dictionary file with an extension .mdd

Value

For read_rdd(), an RddFile object. For read_mdd(), a MddFile object.

Details

Basically, a RddFile and MddFile object is a data.table with 3 additional attributes:

eplus_version: A numeric_version object. The version of EnergyPlus used during the simulation.

eplus_build: A single string. The build tag of EnergyPlus used during the simulation.

datetime: A DateTime (POSIXct). The time when the simulation started.

Examples

# NOT RUN {
# run simulation and get the err file
idf_name <- "1ZoneUncontrolled.idf"
epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
idf_path <- file.path(eplus_config(8.8)$dir, "ExampleFiles", idf_name)
epw_path <- file.path(eplus_config(8.8)$dir, "WeatherData", epw_name)
job <- eplus_job(idf_path, epw_path)

# read the err file
read_rdd(job$locate_output(".rdd"))
read_mdd(job$locate_output(".mdd"))
# }