Run simulations of EnergyPlus models.

run_idf() is a wrapper of EnergyPlus command line interface which enables to run EnergyPlus model with different options.

run_idf(eplus, model, weather, output_dir = NULL, design_day = FALSE,
  annual = FALSE, expand_obj = TRUE, echo = FALSE)

run_multi(eplus, model, weather, output_dir = NULL, design_day = FALSE,
  annual = FALSE, parallel_backend = future::multiprocess)

Arguments

eplus
An acceptable input of use_eplus() and eplus_config().

model
A path of an EnergyPlus IDF or IMF file.

weather
A path of an EnergyPlus EPW weather file.

output_dir
Output directory path. If NULL, which is default, the directory of input model will be used.

design_day
Force design-day-only simulation. Default: FALSE.

annual
Force design-day-only simulation. Default: FALSE.

expand_obj
Where to run ExpandObject preprocessor before simulation. Default: TRUE.

echo
Only applicable to run_idf. Show EnergyPlus simulation process information to the console. If FALSE, which is default, a processx::process object will be return.

parallel_backend
Acceptable input for future::plan().

Details

run_multi() provides the functionality of running multiple models in parallel.

Behind the scene, run_multi() uses the furrr package which provides mapping functions in parallel using future package. It is suggested to run simulations using EplusJob class, which provides much more detailed controls on the simulation and also methods to extract simulation output.

References

Running EnergyPlus from Command Line (EnergyPlus GitHub Repository)

Examples

# NOT RUN {
run_idf(8.8, "input.idf", "weather.epw")
# }