Run simulations of EnergyPlus models.

run_idf() is a wrapper of EnergyPlus command line interface which enables to run EnergyPlus model with different options.

run_idf(eplus, model, weather, output_dir, design_day = FALSE,
  annual = FALSE, expand_obj = TRUE, echo = TRUE)

run_multi(eplus, model, weather, output_dir, design_day = FALSE,
  annual = FALSE, parallel_backend = future::multiprocess)

Arguments

eplus
An acceptable input of use_eplus() and eplus_config().

model
A path of an EnergyPlus IDF or IMF file.

weather
A path of an EnergyPlus EPW weather file.

output_dir
Output directory path. If NULL, the directory of input model is used.

design_day
Force design-day-only simulation. Default: FALSE.

annual
Force design-day-only simulation. Default: FALSE.

expand_obj
Where to run ExpandObject preprocessor before simulation. Default: TRUE.

echo
Only applicable to run_idf. If TRUE, show EnergyPlus simulation process information to the console. If FALSE, which is default, a processx::process object is returned. Default: TRUE.

parallel_backend
Acceptable input for future::plan(). Default: future::multiprocess, which tries to run all models in parallel.

Details

run_multi() provides the functionality of running multiple models in parallel.

Behind the scene, run_multi() uses the furrr package which provides mapping functions in parallel using future package.

It is suggested to run simulations using EplusJob class, which provides much more detailed controls on the simulation and also methods to extract simulation output.

References

Running EnergyPlus from Command Line (EnergyPlus GitHub Repository)

See also

Examples

idf_path <- system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr")

if (is_avail_eplus(8.8)) {
    # run a single model
    epw_path <- file.path(
        eplus_config(8.8)$dir,
        "WeatherData",
        "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
    )

    run_idf(8.8, idf_path, epw_path, output_dir = tempdir())

    # run multiple model in parallel
    idf_paths <- file.path(eplus_config(8.8)$dir, "ExampleFiles",
        c("1ZoneUncontrolled.idf", "1ZoneUncontrolledFourAlgorithms.idf")
    )
    epw_paths <- rep(epw_path, times = 2L)
    output_dirs <- file.path(tempdir(), tools::file_path_sans_ext(basename(idf_paths)))
    run_multi(8.8, idf_paths, epw_paths, output_dir = output_dirs)
}

#> ExpandObjects Started.
#> No expanded file generated.
#> ExpandObjects Finished. Time:     0.000
#> EnergyPlus Starting
#> EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2018.08.17 19:14
#> Processing Data Dictionary
#> Processing Input File
#> Initializing Simulation
#> Reporting Surfaces
#> Beginning Primary Simulation
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Warming up {19}
#> Warming up {20}
#> Warming up {21}
#> Warming up {22}
#> Starting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Starting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Starting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=01/21
#> Continuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=02/10
#> Continuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/02
#> Continuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/22
#> Continuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=04/11
#> Continuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/01
#> Continuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/21
#> Continuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/10
#> Continuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/30
#> Continuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=07/20
#> Continuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/09
#> Continuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/29
#> Continuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=09/18
#> Continuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/08
#> Continuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/28
#> Continuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=11/17
#> Continuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/07
#> Continuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/27
#> Continuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Writing tabular output file results using comma format.
#> Writing tabular output file results using tab format.
#> Writing tabular output file results using text format.
#> Writing tabular output file results using HTML format.
#> Writing tabular output file results using XML format.
#> Writing final SQL reports
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.27sec
#>  ReadVarsESO program completed successfully.
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.12sec
#>  ReadVarsESO program completed successfully.
#> EnergyPlus Run Time=00hr 00min  2.50sec
#>  Progress:                                                                  100% Progress:                                                                  100% Progress:                                                                  100% Progress:                                                                  100% Progress:                                                                  100% Progress:                                                                  100% Progress:                                                                  100% Progress:                                                                  100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ────────────────────────────────                                 100% Progress: ──────────────────────────────────────────────────────────────── 100%
#> 

#> [[1]]
#> [[1]]$status
#> [1] 0
#> 
#> [[1]]$stdout
#> [1] "ExpandObjects Started.\nNo expanded file generated.\nExpandObjects Finished. Time:     0.004\nEnergyPlus Starting\nEnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2018.08.17 19:14\nProcessing Data Dictionary\nProcessing Input File\nInitializing Simulation\nReporting Surfaces\nBeginning Primary Simulation\nInitializing New Environment Parameters\nWarming up {1}\nWarming up {2}\nWarming up {3}\nWarming up {4}\nWarming up {5}\nWarming up {6}\nWarming up {7}\nWarming up {8}\nWarming up {9}\nWarming up {10}\nWarming up {11}\nWarming up {12}\nWarming up {13}\nWarming up {14}\nWarming up {15}\nWarming up {16}\nWarming up {17}\nWarming up {18}\nWarming up {19}\nWarming up {20}\nWarming up {21}\nWarming up {22}\nStarting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB\nInitializing New Environment Parameters\nWarming up {1}\nWarming up {2}\nWarming up {3}\nWarming up {4}\nWarming up {5}\nWarming up {6}\nWarming up {7}\nWarming up {8}\nWarming up {9}\nWarming up {10}\nWarming up {11}\nWarming up {12}\nWarming up {13}\nWarming up {14}\nWarming up {15}\nWarming up {16}\nStarting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB\nInitializing New Environment Parameters\nWarming up {1}\nWarming up {2}\nWarming up {3}\nWarming up {4}\nWarming up {5}\nWarming up {6}\nWarming up {7}\nWarming up {8}\nWarming up {9}\nWarming up {10}\nWarming up {11}\nWarming up {12}\nWarming up {13}\nWarming up {14}\nWarming up {15}\nWarming up {16}\nWarming up {17}\nWarming up {18}\nStarting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=01/21\nContinuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=02/10\nContinuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=03/02\nContinuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=03/22\nContinuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=04/11\nContinuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=05/01\nContinuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=05/21\nContinuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=06/10\nContinuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=06/30\nContinuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=07/20\nContinuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=08/09\nContinuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=08/29\nContinuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=09/18\nContinuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=10/08\nContinuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=10/28\nContinuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=11/17\nContinuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=12/07\nContinuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nUpdating Shadowing Calculations, Start Date=12/27\nContinuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940\nWriting tabular output file results using comma format.\nWriting tabular output file results using tab format.\nWriting tabular output file results using text format.\nWriting tabular output file results using HTML format.\nWriting tabular output file results using XML format.\nWriting final SQL reports\n ReadVarsESO program starting.\n ReadVars Run Time=00hr 00min  0.48sec\n ReadVarsESO program completed successfully.\n ReadVarsESO program starting.\n ReadVars Run Time=00hr 00min  0.20sec\n ReadVarsESO program completed successfully.\nEnergyPlus Run Time=00hr 00min  3.95sec\n"
#> 
#> [[1]]$stderr
#> [1] ""
#> 
#> [[1]]$timeout
#> [1] FALSE
#> 
#> 
#> [[2]]
#> [[2]]$status
#> [1] 0
#> 
#> [[2]]$stdout
#> [1] "ExpandObjects Started.\nNo expanded file generated.\nExpandObjects Finished. Time:     0.004\nEnergyPlus Starting\nEnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2018.08.17 19:14\nProcessing Data Dictionary\nProcessing Input File\nInitializing Simulation\nReporting Surfaces\nBeginning Primary Simulation\nInitializing New Environment Parameters\nWarming up {1}\nWarming up {2}\nWarming up {3}\nWarming up {4}\nWarming up {5}\nWarming up {6}\nWarming up {7}\nWarming up {8}\nWarming up {9}\nWarming up {10}\nStarting Simulation at 01/14 for JAN\nWriting tabular output file results using comma format.\nWriting tabular output file results using tab format.\nWriting tabular output file results using text format.\nWriting tabular output file results using HTML format.\nWriting tabular output file results using XML format.\nWriting final SQL reports\n ReadVarsESO program starting.\n ReadVars Run Time=00hr 00min  0.03sec\n ReadVarsESO program completed successfully.\n ReadVarsESO program starting.\n ReadVars Run Time=00hr 00min  0.02sec\n ReadVarsESO program completed successfully.\nEnergyPlus Run Time=00hr 00min  7.27sec\n"
#> 
#> [[2]]$stderr
#> [1] ""
#> 
#> [[2]]$timeout
#> [1] FALSE
#> 
#>