Run simulations of EnergyPlus models.

run_idf() is a wrapper of EnergyPlus command line interface which enables to run EnergyPlus model with different options.

run_idf(
  model,
  weather,
  output_dir,
  design_day = FALSE,
  annual = FALSE,
  expand_obj = TRUE,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

run_multi(
  model,
  weather,
  output_dir,
  design_day = FALSE,
  annual = FALSE,
  wait = TRUE,
  echo = TRUE,
  eplus = NULL
)

Arguments

model
A path (for run_idf()) or a vector of paths (for run_multi()) of EnergyPlus IDF or IMF files.

weather
A path (for run_idf()) or a vector of paths (for run_multi()) of EnergyPlus EPW weather files. For run_multi(), weather can also be a single EPW file path. In this case, that weather will be used for all simulations; otherwise, model and weather should have the same length.

output_dir
Output directory path (for rum_idf()) or paths (for run_mult()). If NULL, the directory of input model is used. For run_multi(), output_dir, if not NULL, should have the same length as model. Any duplicated combination of model and output_dir is prohibited.

design_day
Force design-day-only simulation. Default: FALSE.

annual
Force design-day-only simulation. Default: FALSE.

expand_obj
Whether to run ExpandObject preprocessor before simulation. Default: TRUE.

wait
If TRUE, R will hang on and wait all EnergyPlus simulations finish. If FALSE, all EnergyPlus simulations are run in the background. and a processx::process object is returned. Note that, if FALSE, R is not blocked even when echo is TRUE. Default: TRUE.

echo
Only applicable when wait is TRUE. Whether to show standard output and error from EnergyPlus command line interface for run_idf() and simulation status for run_multi().Default: TRUE.

eplus
An acceptable input (for run_idf()) or inputs (for run_multi()) of use_eplus() and eplus_config(). If NULL, which is the default, the version of EnergyPlus to use is determined by the version of input model. For run_multi(), eplus, if not NULL, should have the same length as model.

Value

A list for run_idf(). For rum_multi(), a data.table if wait is TRUE or a process if wait is FALSE.

Details

run_multi() provides the functionality of running multiple models in parallel.

run_idf() and run_multi() currently only support EnergyPlus v8.3 and above. This is because eplusr uses EnergyPlus command line interface which is a new feature as of EnergyPlus v8.3.

For run_idf(), a named list will be returned:

idf: The path of IDF file

epw: The path of EPW file

exit_status: The exit code of the process if it has finished and NULL otherwise. Always being NULL if wait is FALSE, but you can manually get the exit code using the process object, i.e. process$get_exit_status() after simulation completed.

start_time: When the EnergyPlus process started.

end_time: When the EnergyPlus process stopped. All being NULL if wait is FALSE, but you can manually check EnergyPlus stdout to get the simulation time

output_dir: The simulation output directory

energyplus: The path of EnergyPlus executable

stdout: All standard output from EnergyPlus. Always being NULL if wait is FALSE, but you can manually get all standard output using process$get_result(), where process is the processx::process object stored in returned element process.

stderr: All standard error from EnergyPlus. Always being NULL if wait is FALSE, but you can manually get all standard output using process$get_result(), where process is the processx::process object stored in returned element process.

process: A processx::process object of current EnergyPlus simulation

For run_multi(), if wait is TRUE, a data.table contains all data (excluding process) with same column names as above, and also another two columns:

index: The index of simulation

status: The status of simulation. Should be one of below:

"completed": the simulation job is completed successfully.

"failed": the simulation job ended with error.

"terminated": the simulation job started but was terminated.

"cancelled": the simulation job was cancelled, i.e. did not start at all.

For run_multi(), if wait is FALSE, a r_process object of background R process which handles all simulation jobs is returned. You can check if the jobs are completed using $is_alive() and get the final data.table using $get_result().

It is suggested to run simulations using EplusJob class and ParametricJob class, which provide much more detailed controls on the simulation and also methods to extract simulation outputs.

References

Running EnergyPlus from Command Line (EnergyPlus GitHub Repository)

See also

Examples

# \dontrun{
idf_path <- system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr")

if (is_avail_eplus(8.8)) {
    # run a single model
    epw_path <- file.path(
        eplus_config(8.8)$dir,
        "WeatherData",
        "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"
    )

    run_idf(idf_path, epw_path, output_dir = tempdir())

    # run multiple model in parallel
    idf_paths <- file.path(eplus_config(8.8)$dir, "ExampleFiles",
        c("1ZoneUncontrolled.idf", "1ZoneUncontrolledFourAlgorithms.idf")
    )
    epw_paths <- rep(epw_path, times = 2L)
    output_dirs <- file.path(tempdir(), tools::file_path_sans_ext(basename(idf_paths)))
    run_multi(idf_paths, epw_paths, output_dir = output_dirs)
}

#> EnergyPlus Starting
#> EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2020.07.10 09:29
#> Processing Data Dictionary
#> Processing Input File
#> Initializing Simulation
#> Reporting Surfaces
#> Beginning Primary Simulation
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Warming up {19}
#> Warming up {20}
#> Warming up {21}
#> Warming up {22}
#> Starting Simulation at 12/21 for DENVER CENTENNIAL  GOLDEN   N ANN HTG 99% CONDNS DB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Starting Simulation at 07/21 for DENVER CENTENNIAL  GOLDEN   N ANN CLG 1% CONDNS DB=>MWB
#> Initializing New Environment Parameters
#> Warming up {1}
#> Warming up {2}
#> Warming up {3}
#> Warming up {4}
#> Warming up {5}
#> Warming up {6}
#> Warming up {7}
#> Warming up {8}
#> Warming up {9}
#> Warming up {10}
#> Warming up {11}
#> Warming up {12}
#> Warming up {13}
#> Warming up {14}
#> Warming up {15}
#> Warming up {16}
#> Warming up {17}
#> Warming up {18}
#> Starting Simulation at 01/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=01/21
#> Continuing Simulation at 01/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=02/10
#> Continuing Simulation at 02/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/02
#> Continuing Simulation at 03/02 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=03/22
#> Continuing Simulation at 03/22 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=04/11
#> Continuing Simulation at 04/11 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/01
#> Continuing Simulation at 05/01 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=05/21
#> Continuing Simulation at 05/21 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/10
#> Continuing Simulation at 06/10 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=06/30
#> Continuing Simulation at 06/30 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=07/20
#> Continuing Simulation at 07/20 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/09
#> Continuing Simulation at 08/09 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=08/29
#> Continuing Simulation at 08/29 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=09/18
#> Continuing Simulation at 09/18 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/08
#> Continuing Simulation at 10/08 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=10/28
#> Continuing Simulation at 10/28 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=11/17
#> Continuing Simulation at 11/17 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/07
#> Continuing Simulation at 12/07 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Updating Shadowing Calculations, Start Date=12/27
#> Continuing Simulation at 12/27 for San Francisco Intl Ap CA USA TMY3 WMO#=724940
#> Writing tabular output file results using comma format.
#> Writing tabular output file results using tab format.
#> Writing tabular output file results using text format.
#> Writing tabular output file results using HTML format.
#> Writing tabular output file results using XML format.
#> Writing final SQL reports
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.22sec
#>  ReadVarsESO program completed successfully.
#>  ReadVarsESO program starting.
#>  ReadVars Run Time=00hr 00min  0.10sec
#>  ReadVarsESO program completed successfully.
#> EnergyPlus Run Time=00hr 00min  1.63sec
#> 

#> 1|RUNNING    --> [IDF]'1ZoneUncontrolled.idf' + [EPW]'USA_CA_San.Francisco....

#> 2|RUNNING    --> [IDF]'1ZoneUncontrolledFourAlgorithms.idf' + [EPW]'USA_CA_...

#> 1|COMPLETED  --> [IDF]'1ZoneUncontrolled.idf' + [EPW]'USA_CA_San.Francisco....

#> 2|COMPLETED  --> [IDF]'1ZoneUncontrolledFourAlgorithms.idf' + [EPW]'USA_CA_...

#>    index    status
#> 1:     1 completed
#> 2:     2 completed
#>                                                                                      idf
#> 1:               /home/travis/.local/EnergyPlus-8-8-0/ExampleFiles/1ZoneUncontrolled.idf
#> 2: /home/travis/.local/EnergyPlus-8-8-0/ExampleFiles/1ZoneUncontrolledFourAlgorithms.idf
#>                                                                                              epw
#> 1: /home/travis/.local/EnergyPlus-8-8-0/WeatherData/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw
#> 2: /home/travis/.local/EnergyPlus-8-8-0/WeatherData/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw
#>    exit_status          start_time            end_time
#> 1:           0 2020-07-10 09:30:00 2020-07-10 09:30:05
#> 2:           0 2020-07-10 09:30:00 2020-07-10 09:30:10
#>                                               energyplus
#> 1: /home/travis/.local/EnergyPlus-8-8-0/energyplus-8.8.0
#> 2: /home/travis/.local/EnergyPlus-8-8-0/energyplus-8.8.0
#>                                         output_dir
#> 1:               /tmp/RtmpeLRCuX/1ZoneUncontrolled
#> 2: /tmp/RtmpeLRCuX/1ZoneUncontrolledFourAlgorithms
#>                                                                                                                                                                            stdout
#> 1: EnergyPlus Starting,EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2020.07.10 09:30,Processing Data Dictionary,Processing Input File,Initializing Simulation,Reporting Surfaces,...
#> 2: EnergyPlus Starting,EnergyPlus, Version 8.8.0-7c3bbe4830, YMD=2020.07.10 09:30,Processing Data Dictionary,Processing Input File,Initializing Simulation,Reporting Surfaces,...
#>                                stderr
#> 1: EnergyPlus Completed Successfully.
#> 2: EnergyPlus Completed Successfully.

# }