Use a specific EnergyPlus Input Data Dictionary (IDD) file

use_idd() takes a path of an EnergyPlus Input Data Dictionary (IDD) file, usually named "Energy+.idd" and return an Idd object. For details on Idd class, please see idd.

use_idd(idd, download = FALSE)

download_idd(ver = "latest", dir = getwd())

avail_idd()

is_avail_idd(ver)

Arguments

idd
A path to an EnergyPlus Input Data Dictionary (IDD) file, usually named as Energy+.idd or a valid version of IDD, e.g. 8.9, "8.9.0".

download
If "TRUE" and argument idd, the IDD file will be downloaded from EnergyPlus GitHub Repository, and saved to tempdir(). It will be parsed after it is downloaded successfully. It is useful in case where you only want to edit an EnergyPlus Input Data File (IDF) directly but do not want to install whole EnergyPlus software.

ver
A valid EnergyPlus version, e.g., 8.7, "8.7" or "8.7.0". For download_idd, The special value "latest" means the latest version (fetched from GitHub releases). Default: "latest".

dir
A directory to indicate where to save the IDD file. Default: current working directory.

Value

use_idd returns an Idd object; download_idd returns the full path of the downloaded IDD file; avail_idd returns a character vector; is_avail_idd returns a single logical vector.

Details

download_idd() downloads specified version of EnergyPlus IDD file from EnergyPlus GitHub Repository.

avail_idd() returns versions of all cached Idd object.

is_avail_idd() returns TRUE if input version of IDD file has been parsed and cached.

eplusr tries to detect all installed EnergyPlus in default installation locations when loading. If argument idd is a version, eplusr will first try to find the cached Idd object of that version, if possible. If failed, and EnergyPlus of that version is available (see avail_eplus()), the "Energy+.idd" distributed with EnergyPlus will be parsed and stored in eplusr Idd cache.

download_idd() downloads EnergyPlus Input Data Dictionary (IDD) file from EnergyPlus GitHub Repository. It is useful in case where you only want to edit an EnergyPlus Input Data File (IDF) directly but do not want to install whole EnergyPlus software.

See also

Examples

# NOT RUN {
use_idd(8.8)

download_idd()
download_idd(8.8)

is_avail_idd(8.8)
# }